General Interface for Boosted Trees

boost_tree is a way to generate a specification of a model before fitting and allows the model to be created using different packages in R or via Spark. The main arguments for the model are:

mtry: The number of predictors that will be randomly sampled at each split when creating the tree models.

trees: The number of trees contained in the ensemble.

min_n: The minimum number of data points in a node that are required for the node to be split further.

tree_depth: The maximum depth of the tree (i.e. number of splits).

learn_rate: The rate at which the boosting algorithm adapts from iteration-to-iteration.

loss_reduction: The reduction in the loss function required to split further.

sample_size: The amount of data exposed to the fitting routine.

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using the others argument. If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update can be used in lieu of recreating the object from scratch.

boost_tree(mode = "unknown", mtry = NULL, trees = NULL,
  min_n = NULL, tree_depth = NULL, learn_rate = NULL,
  loss_reduction = NULL, sample_size = NULL, others = list(), ...)

# S3 method for boost_tree
update(object, mtry = NULL, trees = NULL,
  min_n = NULL, tree_depth = NULL, learn_rate = NULL,
  loss_reduction = NULL, sample_size = NULL, others = list(),
  fresh = FALSE, ...)

Arguments

mode
A single character string for the type of model. Possible values for this model are "unknown", "regression", or "classification".

mtry
An number for the number (or proportion) of predictors that will be randomly sampled at each split when creating the tree models (xgboost only).

trees
An integer for the number of trees contained in the ensemble.

min_n
An integer for the minimum number of data points in a node that are required for the node to be split further.

tree_depth
An integer for the maximum deopth of the tree (i.e. number of splits) (xgboost only).

learn_rate
A number for the rate at which the boosting algorithm adapts from iteration-to-iteration (xgboost only).

loss_reduction
A number for the reduction in the loss function required to split further (xgboost only).

sample_size
An number for the number (or proportion) of data that is exposed to the fitting routine. For xgboost, the sampling is done at at each iteration while C5.0 samples once during traning.

others
A named list of arguments to be used by the underlying models (e.g., xgboost::xgb.train, etc.). .

...
Used for method consistency. Any arguments passed to the ellipses will result in an error. Use others instead.

object
A boosted tree model specification.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

Value

An updated model specification.

Details

The data given to the function are not saved and are only used to determine the mode of the model. For boost_tree, the possible modes are "regression" and "classification".

The model can be created using the fit() function using the following engines:

R: "xgboost", "C5.0"

Spark: "spark"

Main parameter arguments (and those in others) can avoid evaluation until the underlying function is executed by wrapping the argument in rlang::expr() (e.g. mtry = expr(floor(sqrt(p)))).

Engines may have pre-set default arguments when executing the model fit call. These can be changed by using the others argument to pass in the preferred values. For this type of model, the template of the fit calls are:

xgboost classification

xgb_train(x = missing_arg(), y = missing_arg(), nthread = 1, 
    verbose = 0)

xgboost regression

xgb_train(x = missing_arg(), y = missing_arg(), nthread = 1, 
    verbose = 0)

C5.0 classification

C5.0_train(x = missing_arg(), y = missing_arg(), weights = missing_arg())

spark classification

sparklyr::ml_gradient_boosted_trees(x = missing_arg(), formula = missing_arg(), 
    type = "classification", seed = sample.int(10^5, 1))

spark regression

sparklyr::ml_gradient_boosted_trees(x = missing_arg(), formula = missing_arg(), 
    type = "regression", seed = sample.int(10^5, 1))

Note

For models created using the spark engine, there are several differences to consider. First, only the formula interface to via fit is available; using fit_xy will generate an error. Second, the predictions will always be in a spark table format. The names will be the same as documented but without the dots. Third, there is no equivalent to factor columns in spark tables so class predictions are returned as character columns. Fourth, to retain the model object for a new R session (via save), the model$fit element of the parsnip object should be serialized via ml_save(object$fit) and separately saved to disk. In a new session, the object can be reloaded and reattached to the parsnip object.

See also

Examples

boost_tree(mode = "classification", trees = 20)

#> Boosted Tree Model Specification (classification)
#> 
#> Main Arguments:
#>   trees = 20
#> 

# Parameters can be represented by a placeholder:
boost_tree(mode = "regression", mtry = varying())

#> Boosted Tree Model Specification (regression)
#> 
#> Main Arguments:
#>   mtry = varying()
#> 

model <- boost_tree(mtry = 10, min_n = 3)
model

#> Boosted Tree Model Specification (unknown)
#> 
#> Main Arguments:
#>   mtry = 10
#>   min_n = 3
#> 

update(model, mtry = 1)

#> Boosted Tree Model Specification (unknown)
#> 
#> Main Arguments:
#>   mtry = 1
#>   min_n = 3
#> 

update(model, mtry = 1, fresh = TRUE)

#> Boosted Tree Model Specification (unknown)
#> 
#> Main Arguments:
#>   mtry = 1
#>