Tools to Register Models

These functions are similar to constructors and can be used to validate that there are no conflicts with the underlying model structures used by the package.

pred_types

check_mod_val(model, new = FALSE, existence = FALSE)

get_model_env()

check_mode_val(mode)

check_engine_val(eng)

check_arg_val(arg)

check_submodels_val(has_submodel)

check_func_val(func)

check_fit_info(fit_obj)

check_pred_info(pred_obj, type)

check_pkg_val(pkg)

set_new_model(model)

set_model_mode(model, mode)

set_model_engine(model, mode, eng)

set_model_arg(model, eng, parsnip, original, func, has_submodel)

set_dependency(model, eng, pkg)

get_dependency(model)

set_fit(model, mode, eng, value)

get_fit(model)

set_pred(model, mode, eng, type, value)

get_pred_type(model, type)

show_model_info(model)

get_from_env(items)

Arguments

model
A single character string for the model type (e.g. "rand_forest", etc).

new
A single logical to check to see if the model that you are check has not already been registered.

existence
A single logical to check to see if the model has already been registered.

mode
A single character string for the model mode (e.g. "regression").

eng
A single character string for the model engine.

arg
A single character string for the model argument name.

has_submodel
A single logical for whether the argument can make predictions on mutiple submodels at once.

func
A named character vector that describes how to call a function. func should have elements pkg and fun. The former is optional but is recommended and the latter is required. For example, c(pkg = "stats", fun = "lm") would be used to invoke the usual linear regression function. In some cases, it is helpful to use c(fun = "predict") when using a package's predict method.

fit_obj
A list with elements interface, protect, func and defaults. See the package vignette "Making a parsnip model from scratch".

pred_obj
A list with elements pre, post, func, and args. See the package vignette "Making a parsnip model from scratch".

type
A single character value for the type of prediction. Possible values are: 'raw', 'numeric', 'class', 'link', 'prob', 'conf_int', 'pred_int', 'quantile' .

pkg
An options character string for a package name.

parsnip
A single character string for the "harmonized" argument name that parsnip exposes.

original
A single character string for the argument name that underlying model function uses.

value
A list that conforms to the fit_obj or pred_obj description above, depending on context.

items
A character string of objects in the model environment.

Format

An object of class character of length 8.

Details

These functions are available for users to add their own models or engines (in package or otherwise) so that they can be accessed using parsnip. This are more thoroughly documented on the package web site (see references below).

In short, parsnip stores an environment object that contains all of the information and code about how models are used (e.g. fitting, predicting, etc). These functions can be used to add models to that environment as well as helper functions that can be used to makes sure that the model data is in the right format.

References

"Making a parsnip model from scratch" https://tidymodels.github.io/parsnip/articles/articles/Scratch.html

Examples

# Show the information about a model:
show_model_info("rand_forest")

#> Information for `rand_forest`
#>  modes: unknown, classification, regression 
#> 
#>  engines: 
#>    classification: randomForest, ranger, spark
#>    regression:     randomForest, ranger, spark
#> 
#>  arguments: 
#>    ranger:       
#>       mtry  --> mtry
#>       trees --> num.trees
#>       min_n --> min.node.size
#>    randomForest: 
#>       mtry  --> mtry
#>       trees --> ntree
#>       min_n --> nodesize
#>    spark:        
#>       mtry  --> feature_subset_strategy
#>       trees --> num_trees
#>       min_n --> min_instances_per_node
#> 
#>  fit modules:
#>          engine           mode
#>          ranger classification
#>          ranger     regression
#>    randomForest classification
#>    randomForest     regression
#>           spark classification
#>           spark     regression
#> 
#>  prediction modules:
#>              mode       engine                    methods
#>    classification randomForest           class, prob, raw
#>    classification       ranger class, conf_int, prob, raw
#>    classification        spark                class, prob
#>        regression randomForest               numeric, raw
#>        regression       ranger     conf_int, numeric, raw
#>        regression        spark                    numeric
#> 

# Access the model data:
current_code <- get_model_env()
ls(envir = current_code)

#>  [1] "boost_tree"               "boost_tree_args"         
#>  [3] "boost_tree_fit"           "boost_tree_modes"        
#>  [5] "boost_tree_pkgs"          "boost_tree_predict"      
#>  [7] "decision_tree"            "decision_tree_args"      
#>  [9] "decision_tree_fit"        "decision_tree_modes"     
#> [11] "decision_tree_pkgs"       "decision_tree_predict"   
#> [13] "linear_reg"               "linear_reg_args"         
#> [15] "linear_reg_fit"           "linear_reg_modes"        
#> [17] "linear_reg_pkgs"          "linear_reg_predict"      
#> [19] "logistic_reg"             "logistic_reg_args"       
#> [21] "logistic_reg_fit"         "logistic_reg_modes"      
#> [23] "logistic_reg_pkgs"        "logistic_reg_predict"    
#> [25] "mars"                     "mars_args"               
#> [27] "mars_fit"                 "mars_modes"              
#> [29] "mars_pkgs"                "mars_predict"            
#> [31] "mlp"                      "mlp_args"                
#> [33] "mlp_fit"                  "mlp_modes"               
#> [35] "mlp_pkgs"                 "mlp_predict"             
#> [37] "models"                   "modes"                   
#> [39] "multinom_reg"             "multinom_reg_args"       
#> [41] "multinom_reg_fit"         "multinom_reg_modes"      
#> [43] "multinom_reg_pkgs"        "multinom_reg_predict"    
#> [45] "nearest_neighbor"         "nearest_neighbor_args"   
#> [47] "nearest_neighbor_fit"     "nearest_neighbor_modes"  
#> [49] "nearest_neighbor_pkgs"    "nearest_neighbor_predict"
#> [51] "null_model"               "null_model_args"         
#> [53] "null_model_fit"           "null_model_modes"        
#> [55] "null_model_pkgs"          "null_model_predict"      
#> [57] "rand_forest"              "rand_forest_args"        
#> [59] "rand_forest_fit"          "rand_forest_modes"       
#> [61] "rand_forest_pkgs"         "rand_forest_predict"     
#> [63] "surv_reg"                 "surv_reg_args"           
#> [65] "surv_reg_fit"             "surv_reg_modes"          
#> [67] "surv_reg_pkgs"            "surv_reg_predict"        
#> [69] "svm_poly"                 "svm_poly_args"           
#> [71] "svm_poly_fit"             "svm_poly_modes"          
#> [73] "svm_poly_pkgs"            "svm_poly_predict"        
#> [75] "svm_rbf"                  "svm_rbf_args"            
#> [77] "svm_rbf_fit"              "svm_rbf_modes"           
#> [79] "svm_rbf_pkgs"             "svm_rbf_predict"