Tools for documenting packages

These are functions used to create dynamic documentation in Rd files based on which parsnip-related packages are loaded by the user.

These functions can be used to make dynamic lists of documentation help files. parsnip uses these along with files in man/rmd which contain expanded documentation for specific model/engine combinations. find_engine_files() looks for files that have the pattern details_{model}_{engine}.Rd to link to. These files are generated by files named man/rmd/{model}_{engine}.Rmd. make_engine_list() creates a list seen at the top of the model Rd files while make_seealso_list() populates the list seen in "See Also" below. See the details section.

Usage

find_engine_files(mod)

make_engine_list(mod)

make_seealso_list(mod, pkg = "parsnip")

Arguments

mod

A character string for the model file (e.g. "linear_reg")

pkg

A character string for the package where the function is invoked.

Value

make_engine_list() returns a character string that creates a bulleted list of links to more specific help files. make_seealso_list() returns a formatted character string of links. find_engine_files() returns a tibble.

Details

The parsnip documentation is generated dynamically. Part of the Rd file populates a list of engines that depends on what packages are loaded at the time that the man file is loaded. For example, if another package has a new engine for linear_reg(), the parsnip::linear_reg() help can show a link to a detailed help page in the other package.

The process for a package developer to create parsnip documentation is:

Create an engine-specific R file in the R directory with the name {model}_{engine}.R (e.g. boost_tree_C5.0.R). This has a small amount of documentation, as well as the directives "@name details_{model}_{engine}" and "@includeRmd man/rmd/{model}_{engine}.md details".

Copy the file in parsnip that is in man/rmd/aaa.Rmd and put it in the same place in your package.

Write your own man/rmd/{model}_{engine}.Rmd file. This can include packages that are not listed in the DESCRIPTION file. Those are only required when the documentation file is created locally (probably using devtools::document()).

Run devtools::document() so that the .md content is included in the Rd file.

The examples in parsnip can provide guidance for how to organize technical information about the models.

Examples

find_engine_files("linear_reg")
#> # A tibble: 12 × 4
#>    engine     topic                                 default .order
#>    <chr>      <chr>                                 <chr>    <dbl>
#>  1 lm         parsnip:details_linear_reg_lm         "¹"          1
#>  2 brulee     parsnip:details_linear_reg_brulee     ""           2
#>  3 gee        parsnip:details_linear_reg_gee        ""           3
#>  4 glm        parsnip:details_linear_reg_glm        ""           4
#>  5 glmnet     parsnip:details_linear_reg_glmnet     ""           5
#>  6 gls        parsnip:details_linear_reg_gls        ""           6
#>  7 keras      parsnip:details_linear_reg_keras      ""           7
#>  8 lme        parsnip:details_linear_reg_lme        ""           8
#>  9 lmer       parsnip:details_linear_reg_lmer       ""           9
#> 10 spark      parsnip:details_linear_reg_spark      ""          10
#> 11 stan       parsnip:details_linear_reg_stan       ""          11
#> 12 stan_glmer parsnip:details_linear_reg_stan_glmer ""          12
cat(make_engine_list("linear_reg"))
#> There are different ways to fit this model, and the method of estimation is  chosen by setting the model \emph{engine}. The engine-specific pages  for this model are listed  below.
#> 
#> 
#> \itemize{
#>   \item \code{\link[parsnip:details_linear_reg_lm]{lm}¹}
#>   \item \code{\link[parsnip:details_linear_reg_brulee]{brulee}}
#>   \item \code{\link[parsnip:details_linear_reg_gee]{gee}²}
#>   \item \code{\link[parsnip:details_linear_reg_glm]{glm}}
#>   \item \code{\link[parsnip:details_linear_reg_glmnet]{glmnet}}
#>   \item \code{\link[parsnip:details_linear_reg_gls]{gls}²}
#>   \item \code{\link[parsnip:details_linear_reg_keras]{keras}}
#>   \item \code{\link[parsnip:details_linear_reg_lme]{lme}²}
#>   \item \code{\link[parsnip:details_linear_reg_lmer]{lmer}²}
#>   \item \code{\link[parsnip:details_linear_reg_spark]{spark}}
#>   \item \code{\link[parsnip:details_linear_reg_stan]{stan}}
#>   \item \code{\link[parsnip:details_linear_reg_stan_glmer]{stan_glmer}²}
#> }
#> 
#>  
#> ¹ The default engine. ² Requires a parsnip extension package.