Generalized additive models (GAMs)

gen_additive_mod() defines a model that can use smoothed functions of numeric predictors in a generalized linear model.

There are different ways to fit this model. See the engine-specific pages for more details

mgcv (default)

More information on how parsnip is used for modeling is at https://www.tidymodels.org/.

gen_additive_mod(
  mode = "unknown",
  select_features = NULL,
  adjust_deg_free = NULL,
  engine = "mgcv"
)

Arguments

mode
A single character string for the prediction outcome mode. Possible values for this model are "unknown", "regression", or "classification".

select_features
TRUE or FALSE. If TRUE, the model has the ability to eliminate a predictor (via penalization). Increasing adjust_deg_free will increase the likelihood of removing predictors.

adjust_deg_free
If select_features = TRUE, then acts as a multiplier for smoothness. Increase this beyond 1 to produce smoother models.

engine
A single character string specifying what computational engine to use for fitting.

Details

This function only defines what type of model is being fit. Once an engine is specified, the method to fit the model is also defined.

The model is not trained or fit until the fit.model_spec() function is used with the data.

References

https://www.tidymodels.org, Tidy Models with R

See also

Examples

show_engines("gen_additive_mod")

#> # A tibble: 2 x 2
#>   engine mode          
#>   <chr>  <chr>         
#> 1 mgcv   regression    
#> 2 mgcv   classification

gen_additive_mod()

#> GAM Specification (unknown)
#> 
#> Computational engine: mgcv 
#>