Loan Data

Loan Data

Source

Lending Club Statistics https://www.lendingclub.com/info/download-data.action

Value

lending_club

a data frame

Details

These data were downloaded from the Lending Club access site (see below) and are from the first quarter of 2016. A subset of the rows and variables are included here. The outcome is in the variable Class and is either "good" (meaning that the loan was fully paid back or currently on-time) or "bad" (charged off, defaulted, of 21-120 days late). A data dictionary can be found on the source website.

Examples

data(lending_club)
str(lending_club)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	9857 obs. of  23 variables:
#>  $ funded_amnt               : int  16100 32000 10000 16800 3500 10000 11000 15000 6000 20000 ...
#>  $ term                      : Factor w/ 2 levels "term_36","term_60": 1 2 1 2 1 1 1 1 1 2 ...
#>  $ int_rate                  : num  13.99 11.99 16.29 13.67 7.39 ...
#>  $ sub_grade                 : Factor w/ 35 levels "A1","A2","A3",..: 14 11 16 13 4 10 1 7 8 12 ...
#>  $ addr_state                : Factor w/ 50 levels "AK","AL","AR",..: 7 23 35 33 5 43 17 24 34 11 ...
#>  $ verification_status       : Factor w/ 3 levels "Not_Verified",..: 1 3 2 3 2 2 1 1 2 1 ...
#>  $ annual_inc                : num  35000 72000 72000 101000 50100 32000 65000 188000 89000 48000 ...
#>  $ emp_length                : Factor w/ 12 levels "emp_lt_1","emp_1",..: 7 3 3 1 12 1 6 8 3 3 ...
#>  $ delinq_2yrs               : int  0 0 0 0 0 0 0 2 0 1 ...
#>  $ inq_last_6mths            : int  0 0 2 0 0 0 0 0 2 0 ...
#>  $ revol_util                : num  67.4 82.2 74.2 64 78.3 68.1 54.8 33.3 72.3 36 ...
#>  $ acc_now_delinq            : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ open_il_6m                : int  1 5 3 6 0 1 2 8 1 0 ...
#>  $ open_il_12m               : int  0 1 2 1 0 0 0 2 0 0 ...
#>  $ open_il_24m               : int  0 3 3 2 0 0 0 5 0 0 ...
#>  $ total_bal_il              : int  1099 49187 33378 55445 0 7574 20998 170814 19616 0 ...
#>  $ all_util                  : int  48 77 79 64 78 66 54 80 75 36 ...
#>  $ inq_fi                    : int  0 0 1 1 0 0 0 0 2 1 ...
#>  $ inq_last_12m              : int  3 0 4 4 0 0 0 4 6 1 ...
#>  $ delinq_amnt               : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ num_il_tl                 : int  3 9 9 10 3 9 7 21 7 1 ...
#>  $ total_il_high_credit_limit: int  13193 65945 39387 60188 0 12131 39486 177477 26129 0 ...
#>  $ Class                     : Factor w/ 2 levels "bad","good": 2 2 2 2 2 2 2 2 2 2 ...