Model Fit Object Information

An object with class "model_fit" is a container for information about a model that has been fit to the data.

Details

The main elements of the object are:

lvl: A vector of factor levels when the outcome is is a factor. This is NULL when the outcome is not a factor vector.

spec: A model_spec object.

fit: The object produced by the fitting function.

preproc: This contains any data-specific information required to process new a sample point for prediction. For example, if the underlying model function requires arguments x and y and the user passed a formula to fit, the preproc object would contain items such as the terms object and so on. When no information is required, this is NA.

This class and structure is the basis for how parsnip stores model objects after to seeing the data and applying a model.