Model predictions across many sub-models

For some models, predictions can be made on sub-models in the model object.

multi_predict(object, ...)

# S3 method for default
multi_predict(object, ...)

# S3 method for `_xgb.Booster`
multi_predict(object, new_data, type = NULL, trees = NULL, ...)

# S3 method for `_C5.0`
multi_predict(object, new_data, type = NULL, trees = NULL, ...)

# S3 method for `_elnet`
multi_predict(object, new_data, type = NULL, penalty = NULL, ...)

# S3 method for `_lognet`
multi_predict(object, new_data, type = NULL, penalty = NULL, ...)

# S3 method for `_earth`
multi_predict(object, new_data, type = NULL, num_terms = NULL, ...)

# S3 method for `_multnet`
multi_predict(object, new_data, type = NULL, penalty = NULL, ...)

# S3 method for `_train.kknn`
multi_predict(object, new_data, type = NULL, neighbors = NULL, ...)

Arguments

object
A model_fit object.

...
Optional arguments to pass to predict.model_fit(type = "raw") such as type.

new_data
A rectangular data object, such as a data frame.

type
A single character value or NULL. Possible values are "numeric", "class", "prob", "conf_int", "pred_int", "quantile", or "raw". When NULL, predict() will choose an appropriate value based on the model's mode.

trees
An integer vector for the number of trees in the ensemble.

penalty
A numeric vector of penalty values.

num_terms
An integer vector for the number of MARS terms to retain.

neighbors
An integer vector for the number of nearest neighbors.

Value

A tibble with the same number of rows as the data being predicted. There is a list-column named .pred that contains tibbles with multiple rows per sub-model. Note that, within the tibbles, the column names follow the usual standard based on prediction type (i.e. .pred_class for type = "class" and so on).