General Interface for Multinomial Regression Models

multinom_reg is a way to generate a specification of a model before fitting and allows the model to be created using different packages in R, keras, or Spark. The main arguments for the model are:

penalty: The total amount of regularization in the model. Note that this must be zero for some engines.

mixture: The proportion of L1 regularization in the model. Note that this will be ignored for some engines.

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using set_engine. If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update can be used in lieu of recreating the object from scratch.

multinom_reg(mode = "classification", penalty = NULL, mixture = NULL)

# S3 method for multinom_reg
update(object, penalty = NULL, mixture = NULL,
  fresh = FALSE, ...)

Arguments

mode
A single character string for the type of model. The only possible value for this model is "classification".

penalty
An non-negative number representing the total amount of regularization (glmnet, keras, and spark only). For keras models, this corresponds to purely L2 regularization (aka weight decay) while the other models can be a combination of L1 and L2 (depending on the value of mixture).

mixture
A number between zero and one (inclusive) that represents the proportion of regularization that is used for the L2 penalty (i.e. weight decay, or ridge regression) versus L1 (the lasso) (glmnet only).

object
A multinomial regression model specification.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

...
Not used for update.

Details

For multinom_reg, the mode will always be "classification".

The model can be created using the fit() function using the following engines:

R: "glmnet"

Stan: "stan"

keras: "keras"

Note

For models created using the spark engine, there are several differences to consider. First, only the formula interface to via fit is available; using fit_xy will generate an error. Second, the predictions will always be in a spark table format. The names will be the same as documented but without the dots. Third, there is no equivalent to factor columns in spark tables so class predictions are returned as character columns. Fourth, to retain the model object for a new R session (via save), the model$fit element of the parsnip object should be serialized via ml_save(object$fit) and separately saved to disk. In a new session, the object can be reloaded and reattached to the parsnip object.

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are:

glmnet

glmnet::glmnet(x = missing_arg(), y = missing_arg(), weights = missing_arg(), 
    family = "multinomial")

spark

sparklyr::ml_logistic_regression(x = missing_arg(), formula = missing_arg(), 
    weight_col = missing_arg(), family = "multinomial")

keras

parsnip::keras_mlp(x = missing_arg(), y = missing_arg(), hidden_units = 1, 
    act = "linear")

When using glmnet models, there is the option to pass multiple values (or no values) to the penalty argument. This can have an effect on the model object results. When using the predict method in these cases, the return object type depends on the value of penalty. If a single value is given, the results will be a simple numeric vector. When multiple values or no values for penalty are used in multinom_reg, the predict method will return a data frame with columns values and lambda.

See also

Examples

multinom_reg()

#> Multinomial Regression Model Specification (classification)
#> 

# Parameters can be represented by a placeholder:
multinom_reg(penalty = varying())

#> Multinomial Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = varying()
#> 

model <- multinom_reg(penalty = 10, mixture = 0.1)
model

#> Multinomial Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = 10
#>   mixture = 0.1
#> 

update(model, penalty = 1)

#> Multinomial Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = 1
#>   mixture = 0.1
#> 

update(model, penalty = 1, fresh = TRUE)

#> Multinomial Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = 1
#>