Other predict methods.

These are internal functions not meant to be directly called by the user.

# S3 method for model_fit
predict_class(object, new_data, ...)

predict_class(object, ...)

# S3 method for model_fit
predict_classprob(object, new_data, ...)

predict_classprob(object, ...)

# S3 method for model_fit
predict_confint(object, new_data, level = 0.95,
  std_error = FALSE, ...)

predict_confint(object, ...)

# S3 method for model_fit
predict_predint(object, new_data, level = 0.95,
  std_error = FALSE, ...)

predict_predint(object, ...)

# S3 method for model_fit
predict_numeric(object, new_data, ...)

predict_numeric(object, ...)

# S3 method for model_fit
predict_quantile(object, new_data,
  quantile = (1:9)/10, ...)

predict_quantile(object, ...)

Arguments

object
An object of class model_fit

new_data
A rectangular data object, such as a data frame.

...
Ignored. To pass arguments to pass to the underlying function when predict.model_fit(type = "raw"), use the opts argument.

level
A single numeric value between zero and one for the interval estimates.

std_error
A single logical for wether the standard error should be returned (assuming that the model can compute it).

quant
A vector of numbers between 0 and 1 for the quantile being predicted.