Model predictions

Apply a model to create different types of predictions. predict can be used for all types of models and used the "type" argument for more specificity.

# S3 method for model_fit
predict(object, new_data, type = NULL,
  opts = list(), ...)

# S3 method for model_fit
predict_raw(object, new_data, opts = list(), ...)

predict_raw(object, ...)

Arguments

object
An object of class model_fit

new_data
A rectangular data object, such as a data frame.

type
A single character value or NULL. Possible values are "numeric", "class", "probs", "conf_int", "pred_int", "quantile", or "raw". When NULL, predict will choose an appropriate value based on the model's mode.

opts
A list of optional arguments to the underlying predict function that will be used when type = "raw". The list should not include options for the model object or the new data being predicted.

...
Ignored. To pass arguments to pass to the underlying function when predict.model_fit(type = "raw"), use the opts argument.

Value

With the exception of type = "raw", the results of predict.model_fit will be a tibble as many rows in the output as there are rows in new_data and the column names will be predictable.

For numeric results with a single outcome, the tibble will have a .pred column and .pred_Yname for multivariate results.

For hard class predictions, the column is named .pred_class and, when type = "prob", the columns are .pred_classlevel.

type = "conf_int" and type = "pred_int" return tibbles with columns .pred_lower and .pred_upper with an attribute for the confidence level. In the case where intervals can be produces for class probabilities (or other non-scalar outputs), the columns will be named .pred_lower_classlevel and so on.

Quantile predictions return a tibble with a column .pred, which is a list-column. Each list element contains a tibble with columns .pred and .quantile (and perhaps other columns).

Using type = "raw" with predict.model_fit (or using predict_raw) will return the unadulterated results of the prediction function.

In the case of Spark-based models, since table columns cannot contain dots, the same convention is used except 1) no dots appear in names and 2) vectors are never returned but type-specific prediction functions.

Details

If "type" is not supplied to predict, then a choice is made (type = "numeric" for regression models and type = "class" for classification).

predict is designed to provide a tidy result (see "Value" section below) in a tibble output format.

When using type = "conf_int" and type = "pred_int", the options level and std_error can be used. The latter is a logical for an extra column of standard error values (if available).

Examples

library(dplyr)

lm_model <-
  linear_reg() %>%
  fit(mpg ~ ., data = mtcars %>% slice(11:32), engine = "lm")

#> Error in if (object$engine == "spark" && !inherits(eval_env$data, "tbl_spark")) stop("spark objects can only be used with the formula interface to `fit` ",     "with a spark data object.", call. = FALSE): missing value where TRUE/FALSE needed

pred_cars <-
  mtcars %>%
  slice(1:10) %>%
  select(-mpg)

predict(lm_model, pred_cars)

#> Error in predict(lm_model, pred_cars): object 'lm_model' not found

predict(
  lm_model,
  pred_cars,
  type = "conf_int",
  level = 0.90
)

#> Error in predict(lm_model, pred_cars, type = "conf_int", level = 0.9): object 'lm_model' not found

predict(
  lm_model,
  pred_cars,
  type = "raw",
  opts = list(type = "terms")
)

#> Error in predict(lm_model, pred_cars, type = "raw", opts = list(type = "terms")): object 'lm_model' not found