General Interface for Parametric Survival Models

surv_reg is a way to generate a specification of a model before fitting and allows the model to be created using R. The main argument for the model is:

dist: The probability distribution of the outcome.

This argument is converted to its specific names at the time that the model is fit. Other options and argument can be set using the others argument. If left to its default here (NULL), the value is taken from the underlying model functions.

If parameters need to be modified, this function can be used in lieu of recreating the object from scratch.

surv_reg(mode = "regression", dist = NULL, others = list(), ...)

# S3 method for surv_reg
update(object, dist = NULL, others = list(),
  fresh = FALSE, ...)

Arguments

mode
A single character string for the type of model. The only possible value for this model is "regression".

dist
A character string for the outcome distribution. "weibull" is the default.

others
A named list of arguments to be used by the underlying models (e.g., flexsurv::flexsurvreg). These are not evaluated until the model is fit and will be substituted into the model fit expression.

...
Used for S3 method consistency. Any arguments passed to the ellipses will result in an error. Use others instead.

object
A survival regression model specification.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

Value

An updated model specification.

Details

The data given to the function are not saved and are only used to determine the mode of the model. For surv_reg,the mode will always be "regression".

Since survival models typically involve censoring (and require the use of survival::Surv() objects), the fit() function will require that the survival model be specified via the formula or recipes interfaces.

For recipes, right censoring indicators should be specified using the "censoring var" role (see examples below). Also, for the engine that uses flexsurv::flexsurvfit, extra roles can be used for non-location parameters (e.g. sigma or sdlog) so that other distributional parameters can be functions or covariates. See the example below as well as Jackson (2016).

Also, for the flexsurv::flexsurvfit engine, the typical strata function cannot be used. To achieve the same effect, the extra parameter roles can be used (as described above).

The model can be created using the fit() function using the following engines:

R: "flexsurv"

References

Jackson, C. (2016). flexsurv: A Platform for Parametric Survival Modeling in R. Journal of Statistical Software, 70(8), 1 - 33.

See also

Examples

surv_reg()

#> Parametric Survival Regression Model Specification (regression)
#> 

# Parameters can be represented by a placeholder:
surv_reg(dist = varying())

#> Parametric Survival Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   dist = varying()
#> 

# Examples of using recipes with the `fit` function

library(dplyr)
library(recipes)
library(survival)
data(lung)

surv_rec <- recipe(time ~ ., data = lung) %>%
  add_role(status, new_role = "censoring var") %>%
  # exclude some vars from being in the model
  add_role(inst, sex, ph.karno, pat.karno, meal.cal, wt.loss,
           new_role = "other variables")

#> Warning: Changing role(s) for status

#> Warning: Changing role(s) for inst, sex, ph.karno, pat.karno, meal.cal, wt.loss

log_normal_mod <- surv_reg(dist = "lnorm")

fit(log_normal_mod, recipe = surv_rec, data = lung, engine = "flexsurv")

#> Call:
#> flexsurv::flexsurvreg(formula = Surv(time, status) ~ age + ph.ecog, 
#>     data = data, dist = "lnorm")
#> 
#> Estimates: 
#>          data mean  est       L95%      U95%      se        exp(est)  L95%    
#> meanlog        NA    7.41390   6.37130   8.45651   0.53195        NA        NA
#> sdlog          NA    1.05052   0.94101   1.17278   0.05901        NA        NA
#> age      62.45815   -0.02253  -0.03906  -0.00599   0.00844   0.97772   0.96169
#> ph.ecog   0.95154   -0.34580  -0.55109  -0.14051   0.10474   0.70765   0.57632
#>          U95%    
#> meanlog        NA
#> sdlog          NA
#> age       0.99403
#> ph.ecog   0.86891
#> 
#> N = 227,  Events: 164,  Censored: 63
#> Total time at risk: 69522
#> Log-likelihood = -1152.728, df = 4
#> AIC = 2313.455
#> 

# make the normal variance be a function of gender:

strata_model <- surv_rec %>%
  add_role(sex, new_role = "sdlog")

#> Warning: Changing role(s) for sex

fit(log_normal_mod, recipe = strata_model, data = lung, engine = "flexsurv")

#> Call:
#> flexsurv::flexsurvreg(formula = Surv(time, status) ~ age + ph.ecog + 
#>     sdlog(sex), data = data, dist = "lnorm")
#> 
#> Estimates: 
#>             data mean  est       L95%      U95%      se        exp(est)
#> meanlog           NA    7.35177   6.29235   8.41119   0.54053        NA
#> sdlog             NA    1.15480   0.81414   1.63800   0.20595        NA
#> age         62.45815   -0.02144  -0.03837  -0.00452   0.00863   0.97878
#> ph.ecog      0.95154   -0.33938  -0.54554  -0.13322   0.10519   0.71221
#> sdlog(sex)   1.39648   -0.07113  -0.31799   0.17572   0.12595   0.93134
#>             L95%      U95%    
#> meanlog           NA        NA
#> sdlog             NA        NA
#> age          0.96236   0.99549
#> ph.ecog      0.57953   0.87527
#> sdlog(sex)   0.72761   1.19210
#> 
#> N = 227,  Events: 164,  Censored: 63
#> Total time at risk: 69522
#> Log-likelihood = -1152.571, df = 5
#> AIC = 2315.142
#> 

model <- surv_reg(dist = "weibull")
model

#> Parametric Survival Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   dist = weibull
#> 

update(model, dist = "lnorm")

#> Parametric Survival Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   dist = lnorm
#>