Parametric survival regression

survival_reg() defines a parametric survival model. This function can fit censored regression models.

There are different ways to fit this model, and the method of estimation is chosen by setting the model engine. The engine-specific pages for this model are listed below.

survival¹²

flexsurv²

¹ The default engine. ² Requires a parsnip extension package.
More information on how parsnip is used for modeling is at https://www.tidymodels.org/.

Usage

survival_reg(mode = "censored regression", engine = "survival", dist = NULL)

Arguments

mode

A single character string for the prediction outcome mode. The only possible value for this model is "censored regression".

engine

A single character string specifying what computational engine to use for fitting.

dist

A character string for the probability distribution of the outcome. The default is "weibull".

Details

This function only defines what type of model is being fit. Once an engine is specified, the method to fit the model is also defined. See set_engine() for more on setting the engine, including how to set engine arguments.

The model is not trained or fit until the fit() function is used with the data.

Since survival models typically involve censoring (and require the use of survival::Surv() objects), the fit.model_spec() function will require that the survival model be specified via the formula interface.

References

https://www.tidymodels.org, Tidy Modeling with R, searchable table of parsnip models

See also

Examples

show_engines("survival_reg")
#> # A tibble: 0 × 2
#> # … with 2 variables: engine <chr>, mode <chr>

survival_reg(mode = "censored regression", dist = "weibull")
#> parsnip could not locate an implementation for `survival_reg` censored regression model specifications using the `survival` engine.
#> ℹ The parsnip extension package censored implements support for this specification.
#> ℹ Please install (if needed) and load to continue.
#> Parametric Survival Regression Model Specification (censored regression)
#> 
#> Main Arguments:
#>   dist = weibull
#> 
#> Computational engine: survival 
#>