Linear support vector machines

svm_linear() defines a support vector machine model. For classification, the model tries to maximize the width of the margin between classes. For regression, the model optimizes a robust loss function that is only affected by very large model residuals.

This SVM model uses a linear function to create the decision boundary or regression line.

There are different ways to fit this model. The method of estimation is chosen by setting the model engine.

The engine-specific pages for this model are listed below and contain the details:

LiblineaR (default)

kernlab

More information on how parsnip is used for modeling is at https://www.tidymodels.org/.

svm_linear(mode = "unknown", engine = "LiblineaR", cost = NULL, margin = NULL)

Arguments

mode
A single character string for the prediction outcome mode. Possible values for this model are "unknown", "regression", or "classification".

engine
A single character string specifying what computational engine to use for fitting.

cost
A positive number for the cost of predicting a sample within or on the wrong side of the margin

margin
A positive number for the epsilon in the SVM insensitive loss function (regression only)

Details

This function only defines what type of model is being fit. Once an engine is specified, the method to fit the model is also defined.

The model is not trained or fit until the fit.model_spec() function is used with the data.

References

https://www.tidymodels.org, Tidy Modeling with R

See also

Examples

show_engines("svm_linear")

#> # A tibble: 4 × 2
#>   engine    mode          
#>   <chr>     <chr>         
#> 1 LiblineaR classification
#> 2 LiblineaR regression    
#> 3 kernlab   classification
#> 4 kernlab   regression    

svm_linear(mode = "classification")

#> Linear Support Vector Machine Specification (classification)
#> 
#> Computational engine: LiblineaR 
#>