General interface for radial basis function support vector machines

svm_rbf is a way to generate a specification of a model before fitting and allows the model to be created using different packages in R or via Spark. The main arguments for the model are:

cost: The cost of predicting a sample within or on the wrong side of the margin.

rbf_sigma: The precision parameter for the radial basis function.

margin: The epsilon in the SVM insensitive loss function (regression only)

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using set_engine. If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update can be used in lieu of recreating the object from scratch.

svm_rbf(mode = "unknown", cost = NULL, rbf_sigma = NULL,
  margin = NULL)

# S3 method for svm_rbf
update(object, cost = NULL, rbf_sigma = NULL,
  margin = NULL, fresh = FALSE, ...)

Arguments

mode
A single character string for the type of model. Possible values for this model are "unknown", "regression", or "classification".

cost
A positive number for the cost of predicting a sample within or on the wrong side of the margin

rbf_sigma
A positive number for radial basis function.

margin
A positive number for the epsilon in the SVM insensitive loss function (regression only)

object
A radial basis function SVM model specification.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

...
Not used for update.

Details

The model can be created using the fit() function using the following engines:

R: "kernlab"

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are::

kernlab classification

kernlab::ksvm(x = missing_arg(), y = missing_arg(), kernel = "rbfdot", 
    prob.model = TRUE)

kernlab regression

kernlab::ksvm(x = missing_arg(), y = missing_arg(), kernel = "rbfdot")

See also

Examples

svm_rbf(mode = "classification", rbf_sigma = 0.2)

#> Radial Basis Function Support Vector Machine Specification (classification)
#> 
#> Main Arguments:
#>   rbf_sigma = 0.2
#> 

# Parameters can be represented by a placeholder:
svm_rbf(mode = "regression", cost = varying())

#> Radial Basis Function Support Vector Machine Specification (regression)
#> 
#> Main Arguments:
#>   cost = varying()
#> 

model <- svm_rbf(cost = 10, rbf_sigma = 0.1)
model

#> Radial Basis Function Support Vector Machine Specification (unknown)
#> 
#> Main Arguments:
#>   cost = 10
#>   rbf_sigma = 0.1
#> 

update(model, cost = 1)

#> Radial Basis Function Support Vector Machine Specification (unknown)
#> 
#> Main Arguments:
#>   cost = 1
#>   rbf_sigma = 0.1
#> 

update(model, cost = 1, fresh = TRUE)

#> Radial Basis Function Support Vector Machine Specification (unknown)
#> 
#> Main Arguments:
#>   cost = 1
#>