Update a Random Forest Specification

If parameters need to be modified, this function can be used in lieu of recreating the object from scratch.

# S3 method for rand_forest
update(object, mtry = NULL, trees = NULL,
  min_n = NULL, others = list(), fresh = FALSE, ...)

Arguments

object
A random forest model specification.

mtry
An integer for the number of predictors that will be randomly sampled at each split when creating the tree models.

trees
An integer for the number of trees contained in the ensemble.

min_n
An integer for the minimum number of data points in a node that are required for the node to be split further.

others
A named list of arguments to be used by the underlying models (e.g., ranger::ranger, randomForest::randomForest, etc.). .

fresh
A logical for whether the arguments should be modifed in-place of or replaced wholesale.

...
Used for method consistency. Any arguments passed to the ellipses will result in an error. Use others instead.

Value

An updated model specification.

Examples

model <- rand_forest(mtry = 10, min_n = 3)
model

#> Random Forest Model Specification (unknown)
#> 
#> Main Arguments:
#>   mtry = 10
#>   min_n = 3
#> 
#> 

update(model, mtry = 1)

#> Random Forest Model Specification (unknown)
#> 
#> Main Arguments:
#>   mtry = 1
#>   min_n = 3
#> 
#> 

update(model, mtry = 1, fresh = TRUE)

#> Random Forest Model Specification (unknown)
#> 
#> Main Arguments:
#>   mtry = 1
#> 
#>