Determine varying arguments

varying_args takes a model specification and lists all of the arguments along with whether they are fully specified or not.

varying_args(x, id, ...)

# S3 method for model_spec
varying_args(x, id = NULL, ...)

# S3 method for recipe
varying_args(x, id = NULL, ...)

# S3 method for step
varying_args(x, id = NULL, ...)

Arguments

x
An object

id
A string describing the object x.

...
Not currently used.

Value

A tibble with columns for the parameter name (name), whether is contains any varying value (varying), the id for the object, and the class that was used to call the method (type).

Examples

library(dplyr)
library(rlang)

rand_forest() %>% varying_args(id = "plain")

#> Warning: `list_len()` is soft-deprecated as of rlang 0.2.0.
#> Please use `new_list()` instead
#> This warning is displayed once per session.

#> # A tibble: 3 x 4
#>   name  varying id    type      
#>   <chr> <lgl>   <chr> <chr>     
#> 1 mtry  FALSE   plain model_spec
#> 2 trees FALSE   plain model_spec
#> 3 min_n FALSE   plain model_spec

rand_forest(mtry = varying()) %>% varying_args(id = "one arg")

#> # A tibble: 3 x 4
#>   name  varying id      type      
#>   <chr> <lgl>   <chr>   <chr>     
#> 1 mtry  TRUE    one arg model_spec
#> 2 trees FALSE   one arg model_spec
#> 3 min_n FALSE   one arg model_spec

rand_forest(others = list(sample.fraction = varying())) %>%
  varying_args(id = "only others")

#> # A tibble: 4 x 4
#>   name            varying id          type      
#>   <chr>           <lgl>   <chr>       <chr>     
#> 1 mtry            FALSE   only others model_spec
#> 2 trees           FALSE   only others model_spec
#> 3 min_n           FALSE   only others model_spec
#> 4 sample.fraction TRUE    only others model_spec

rand_forest(
  others = list(
    strata = expr(Class),
    sampsize = c(varying(), varying())
  )
) %>%
  varying_args(id = "add an expr")

#> # A tibble: 5 x 4
#>   name     varying id          type      
#>   <chr>    <lgl>   <chr>       <chr>     
#> 1 mtry     FALSE   add an expr model_spec
#> 2 trees    FALSE   add an expr model_spec
#> 3 min_n    FALSE   add an expr model_spec
#> 4 strata   FALSE   add an expr model_spec
#> 5 sampsize TRUE    add an expr model_spec

rand_forest(others = list(classwt = c(class1 = 1, class2 = varying()))) %>%
  varying_args(id = "list of values")

#> # A tibble: 4 x 4
#>   name    varying id             type      
#>   <chr>   <lgl>   <chr>          <chr>     
#> 1 mtry    FALSE   list of values model_spec
#> 2 trees   FALSE   list of values model_spec
#> 3 min_n   FALSE   list of values model_spec
#> 4 classwt TRUE    list of values model_spec