Boosted trees via C5.0

C5.0_train is a wrapper for the C5.0() function in the C50 package that fits tree-based models where all of the model arguments are in the main function.

Usage

C5.0_train(x, y, weights = NULL, trials = 15, minCases = 2, sample = 0, ...)

Arguments

x

A data frame or matrix of predictors.

y

A factor vector with 2 or more levels

weights

An optional numeric vector of case weights. Note that the data used for the case weights will not be used as a splitting variable in the model (see https://www.rulequest.com/see5-info.html for Quinlan's notes on case weights).

trials

An integer specifying the number of boosting iterations. A value of one indicates that a single model is used.

minCases

An integer for the smallest number of samples that must be put in at least two of the splits.

sample

A value between (0, .999) that specifies the random proportion of the data should be used to train the model. By default, all the samples are used for model training. Samples not used for training are used to evaluate the accuracy of the model in the printed output. A value of zero means that all the training data are used.

...

Other arguments to pass.

Value

A fitted C5.0 model.