Data Set Characteristics Available when Fitting Models

When using the fit functions there are some variables that will be available for use in arguments. For example, if the user would like to choose an argument value based on the current number of rows in a data set, the n_obs variable can be used. See Details below.

Details

Existing variables:

n_obs: the current number of rows in the data set.

n_cols: the number of columns in the data set that are associated with the predictors prior to dummy variable creation.

n_preds: the number of predictors after dummy variables are created (if any).

n_facts: the number of factor predictors in the dat set.

n_levs: If the outcome is a factor, this is a table with the counts for each level (and NA otherwise)

For example, if you use the model formula Sepal.Width ~ . with the iris data, the values would be

 n_cols  =   4     (the 4 columns in `iris`)
 n_preds =   5     (3 numeric columns + 2 from Species dummy variables)
 n_obs   = 150
 n_levs  =  NA     (no factor outcome)
 n_facts =   1     (the Species predictor)

If the formula Species ~ . where used:

 n_cols  =   4     (the 4 numeric columns in `iris`)
 n_preds =   4     (same)
 n_obs   = 150
 n_levs  =  c(setosa = 50, versicolor = 50, virginica = 50)
 n_facts =   0

To use these in a model fit, either expression or rlang::expr can be used to delay the evaluation of the argument value until the time when the model is run via fit (and the variables listed above are available). For example:

library(rlang)
    data("lending_club")
    rand_forest(mode = "classification", mtry = expr(n_cols - 2))

When no instance of expr is found in any of the argument values, the descriptor calculation code will not be executed.