Tools for dynamically documenting packages

These are functions used to create dynamic documentation in Rd files based on which parsnip-related packages are loaded by the user.

These functions can be used to make dynamic lists of documentation help files. parsnip uses these, along with files contained in man/rmd containing expanded documentation, for specific model/engine combinations. find_engine_files() looks for files that have the pattern details_{model}_{engine}.Rd to link to. These files are generated by files named man/rmd/details_{model}_{engine}.Rmd. make_engine_list() creates a list seen at the top of the model Rd files while make_seealso_list() populates the list seen in "See Also" below. See the details section.

find_engine_files(mod, pkg = "parsnip")

make_engine_list(mod, pkg = "parsnip")

make_seealso_list(mod, pkg = "parsnip")

Arguments

mod
A character string for the model file (e.g. "linear_reg")

pkg
A character string for the package where the function is invoked.

Value

make_engine_list() returns a character string that creates a bulleted list of links to more specific help files.

make_seealso_list() returns a formatted character string of links.

find_engine_files() returns a tibble.

Details

The parsnip documentation is generated dynamically. Part of the Rd file populates a list of engines that depends on what packages are loaded at the time that the man file is loaded. For example, if another package has a new engine for linear_reg(), the parsnip::linear_reg() help can show a link to a detailed help page in the other package.

To enable this, the process for a package developer is to:

Create an engine-specific R file in the R directory with the name {model}_{engine}.R (e.g. boost_tree_C5.0.R). This has a small amount of documentation, as well as the directive "@includeRmd man/rmd/{model}_{engine}.Rmd details".

Copy the file in parsnip that is in man/rmd/setup.Rmd and put it in the same place in your package.

Write your own man/rmd/{model}_{engine}.Rmd file. This can include packages that are not listed in the DESCRIPTION file. Those are only required when the documentation file is created locally (probably using devtools::document().

Run devtools::document() so that the Rmd content is included in the Rd file.

The examples in parsnip can provide guidance for how to organize technical information about the models.

Examples

find_engine_files("linear_reg")

#> # A tibble: 5 × 4
#>   engine topic                             .order default     
#>   <chr>  <chr>                              <int> <chr>       
#> 1 lm     parsnip:details_linear_reg_lm          1 " (default)"
#> 2 glmnet parsnip:details_linear_reg_glmnet      2 ""          
#> 3 stan   parsnip:details_linear_reg_stan        3 ""          
#> 4 spark  parsnip:details_linear_reg_spark       4 ""          
#> 5 keras  parsnip:details_linear_reg_keras       5 ""          

cat(make_engine_list("linear_reg"))

#> The engine-specific pages for this model are listed  below and contain the details:
#> 
#> \itemize{
#>   \item \code{\link[parsnip:details_linear_reg_lm]{lm}  (default) }
#>   \item \code{\link[parsnip:details_linear_reg_glmnet]{glmnet}  }
#>   \item \code{\link[parsnip:details_linear_reg_stan]{stan}  }
#>   \item \code{\link[parsnip:details_linear_reg_spark]{spark}  }
#>   \item \code{\link[parsnip:details_linear_reg_keras]{keras}  }
#> }