Generalized additive models (GAMs)

gen_additive_mod() defines a model that can use smoothed functions of numeric predictors in a generalized linear model.

gen_additive_mod(
  mode = "unknown",
  select_features = NULL,
  adjust_deg_free = NULL
)

Arguments

mode
A single character string for the prediction outcome mode. Possible values for this model are "unknown", "regression", or "classification".

select_features
TRUE or FALSE. If this is TRUE then can add an extra penalty to each term so that it can be penalized to zero. This means that the smoothing parameter estimation that is part of fitting can completely remove terms from the model. If the corresponding smoothing parameter is estimated as zero then the extra penalty has no effect. Use adjust_deg_free to increase level of penalization.

adjust_deg_free
If select_features = TRUE, then acts as a multiplier for smoothness. Increase this beyond 1 to produce smoother models.

Value

A parsnip model specification

Details

There are different ways to fit this model. See the engine-specific pages for more details

More information on how parsnip is used for modeling is at https://www.tidymodels.org.

This function only defines what type of model is being fit. Once an engine is specified, the method to fit the model is also defined.

The model is not trained or fit until the fit.model_spec() function is used with the data.

gam

This engine uses mgcv::gam() and has the following parameters, which can be modified through the set_engine() function.

## function (formula, family = gaussian(), data = list(), weights = NULL, 
##     subset = NULL, na.action, offset = NULL, method = "GCV.Cp", optimizer = c("outer", 
##         "newton"), control = list(), scale = 0, select = FALSE, knots = NULL, 
##     sp = NULL, min.sp = NULL, H = NULL, gamma = 1, fit = TRUE, paraPen = NULL, 
##     G = NULL, in.out = NULL, drop.unused.levels = TRUE, drop.intercept = NULL, 
##     discrete = FALSE, ...)

Fit Details

MGCV Formula Interface

Fitting GAMs is accomplished using parameters including:

mgcv::s(): GAM spline smooths

mgcv::te(): GAM tensor product smooths

These are applied in the fit() function:

fit(value ~ s(date_mon, k = 12) + s(date_num), data = df)

References

https://www.tidymodels.org, Tidy Models with R

Examples

#show_engines("gen_additive_mod")

#gen_additive_mod()