Working with the parsnip model environment

These functions read and write to the environment where the package stores information about model specifications.

get_model_env()

get_from_env(items)

set_in_env(...)

set_env_val(name, value)

Arguments

items
A character string of objects in the model environment.

...
Named values that will be assigned to the model environment.

name
A single character value for a new symbol in the model environment.

value
A single value for a new value in the model environment.

References

"Making a parsnip model from scratch" https://tidymodels.github.io/parsnip/articles/articles/Scratch.html

Examples

# Access the model data:
current_code <- get_model_env()
ls(envir = current_code)

#>  [1] "boost_tree"               "boost_tree_args"         
#>  [3] "boost_tree_fit"           "boost_tree_modes"        
#>  [5] "boost_tree_pkgs"          "boost_tree_predict"      
#>  [7] "decision_tree"            "decision_tree_args"      
#>  [9] "decision_tree_fit"        "decision_tree_modes"     
#> [11] "decision_tree_pkgs"       "decision_tree_predict"   
#> [13] "linear_reg"               "linear_reg_args"         
#> [15] "linear_reg_fit"           "linear_reg_modes"        
#> [17] "linear_reg_pkgs"          "linear_reg_predict"      
#> [19] "logistic_reg"             "logistic_reg_args"       
#> [21] "logistic_reg_fit"         "logistic_reg_modes"      
#> [23] "logistic_reg_pkgs"        "logistic_reg_predict"    
#> [25] "mars"                     "mars_args"               
#> [27] "mars_fit"                 "mars_modes"              
#> [29] "mars_pkgs"                "mars_predict"            
#> [31] "mlp"                      "mlp_args"                
#> [33] "mlp_fit"                  "mlp_modes"               
#> [35] "mlp_pkgs"                 "mlp_predict"             
#> [37] "models"                   "modes"                   
#> [39] "multinom_reg"             "multinom_reg_args"       
#> [41] "multinom_reg_fit"         "multinom_reg_modes"      
#> [43] "multinom_reg_pkgs"        "multinom_reg_predict"    
#> [45] "nearest_neighbor"         "nearest_neighbor_args"   
#> [47] "nearest_neighbor_fit"     "nearest_neighbor_modes"  
#> [49] "nearest_neighbor_pkgs"    "nearest_neighbor_predict"
#> [51] "null_model"               "null_model_args"         
#> [53] "null_model_fit"           "null_model_modes"        
#> [55] "null_model_pkgs"          "null_model_predict"      
#> [57] "rand_forest"              "rand_forest_args"        
#> [59] "rand_forest_fit"          "rand_forest_modes"       
#> [61] "rand_forest_pkgs"         "rand_forest_predict"     
#> [63] "surv_reg"                 "surv_reg_args"           
#> [65] "surv_reg_fit"             "surv_reg_modes"          
#> [67] "surv_reg_pkgs"            "surv_reg_predict"        
#> [69] "svm_poly"                 "svm_poly_args"           
#> [71] "svm_poly_fit"             "svm_poly_modes"          
#> [73] "svm_poly_pkgs"            "svm_poly_predict"        
#> [75] "svm_rbf"                  "svm_rbf_args"            
#> [77] "svm_rbf_fit"              "svm_rbf_modes"           
#> [79] "svm_rbf_pkgs"             "svm_rbf_predict"