Working with the parsnip model environment

These functions read and write to the environment where the package stores information about model specifications.

get_model_env()

get_from_env(items)

set_in_env(...)

set_env_val(name, value)

Arguments

items
A character string of objects in the model environment.

...
Named values that will be assigned to the model environment.

name
A single character value for a new symbol in the model environment.

value
A single value for a new value in the model environment.

References

"How to build a parsnip model" https://www.tidymodels.org/learn/develop/models/

Examples

# Access the model data:
current_code <- get_model_env()
ls(envir = current_code)

#>   [1] "bart"                         "bart_args"                   
#>   [3] "bart_encoding"                "bart_fit"                    
#>   [5] "bart_modes"                   "bart_pkgs"                   
#>   [7] "bart_predict"                 "boost_tree"                  
#>   [9] "boost_tree_args"              "boost_tree_encoding"         
#>  [11] "boost_tree_fit"               "boost_tree_modes"            
#>  [13] "boost_tree_pkgs"              "boost_tree_predict"          
#>  [15] "decision_tree"                "decision_tree_args"          
#>  [17] "decision_tree_encoding"       "decision_tree_fit"           
#>  [19] "decision_tree_modes"          "decision_tree_pkgs"          
#>  [21] "decision_tree_predict"        "gen_additive_mod"            
#>  [23] "gen_additive_mod_args"        "gen_additive_mod_encoding"   
#>  [25] "gen_additive_mod_fit"         "gen_additive_mod_modes"      
#>  [27] "gen_additive_mod_pkgs"        "gen_additive_mod_predict"    
#>  [29] "linear_reg"                   "linear_reg_args"             
#>  [31] "linear_reg_encoding"          "linear_reg_fit"              
#>  [33] "linear_reg_modes"             "linear_reg_pkgs"             
#>  [35] "linear_reg_predict"           "logistic_reg"                
#>  [37] "logistic_reg_args"            "logistic_reg_encoding"       
#>  [39] "logistic_reg_fit"             "logistic_reg_modes"          
#>  [41] "logistic_reg_pkgs"            "logistic_reg_predict"        
#>  [43] "mars"                         "mars_args"                   
#>  [45] "mars_encoding"                "mars_fit"                    
#>  [47] "mars_modes"                   "mars_pkgs"                   
#>  [49] "mars_predict"                 "mlp"                         
#>  [51] "mlp_args"                     "mlp_encoding"                
#>  [53] "mlp_fit"                      "mlp_modes"                   
#>  [55] "mlp_pkgs"                     "mlp_predict"                 
#>  [57] "models"                       "modes"                       
#>  [59] "multinom_reg"                 "multinom_reg_args"           
#>  [61] "multinom_reg_encoding"        "multinom_reg_fit"            
#>  [63] "multinom_reg_modes"           "multinom_reg_pkgs"           
#>  [65] "multinom_reg_predict"         "nearest_neighbor"            
#>  [67] "nearest_neighbor_args"        "nearest_neighbor_encoding"   
#>  [69] "nearest_neighbor_fit"         "nearest_neighbor_modes"      
#>  [71] "nearest_neighbor_pkgs"        "nearest_neighbor_predict"    
#>  [73] "null_model"                   "null_model_args"             
#>  [75] "null_model_encoding"          "null_model_fit"              
#>  [77] "null_model_modes"             "null_model_pkgs"             
#>  [79] "null_model_predict"           "proportional_hazards"        
#>  [81] "proportional_hazards_args"    "proportional_hazards_fit"    
#>  [83] "proportional_hazards_modes"   "proportional_hazards_pkgs"   
#>  [85] "proportional_hazards_predict" "rand_forest"                 
#>  [87] "rand_forest_args"             "rand_forest_encoding"        
#>  [89] "rand_forest_fit"              "rand_forest_modes"           
#>  [91] "rand_forest_pkgs"             "rand_forest_predict"         
#>  [93] "surv_reg"                     "surv_reg_args"               
#>  [95] "surv_reg_encoding"            "surv_reg_fit"                
#>  [97] "surv_reg_modes"               "surv_reg_pkgs"               
#>  [99] "surv_reg_predict"             "survival_reg"                
#> [101] "survival_reg_args"            "survival_reg_fit"            
#> [103] "survival_reg_modes"           "survival_reg_pkgs"           
#> [105] "survival_reg_predict"         "svm_linear"                  
#> [107] "svm_linear_args"              "svm_linear_encoding"         
#> [109] "svm_linear_fit"               "svm_linear_modes"            
#> [111] "svm_linear_pkgs"              "svm_linear_predict"          
#> [113] "svm_poly"                     "svm_poly_args"               
#> [115] "svm_poly_encoding"            "svm_poly_fit"                
#> [117] "svm_poly_modes"               "svm_poly_pkgs"               
#> [119] "svm_poly_predict"             "svm_rbf"                     
#> [121] "svm_rbf_args"                 "svm_rbf_encoding"            
#> [123] "svm_rbf_fit"                  "svm_rbf_modes"               
#> [125] "svm_rbf_pkgs"                 "svm_rbf_predict"