Working with the parsnip model environment

These functions read and write to the environment where the package stores information about model specifications.

get_model_env()

get_from_env(items)

set_in_env(...)

set_env_val(name, value)

Arguments

items
A character string of objects in the model environment.

...
Named values that will be assigned to the model environment.

name
A single character value for a new symbol in the model environment.

value
A single value for a new value in the model environment.

References

"How to build a parsnip model" https://www.tidymodels.org/learn/develop/models/

Examples

# Access the model data:
current_code <- get_model_env()
ls(envir = current_code)

#>   [1] "bag_mars"                     "bag_mars_args"               
#>   [3] "bag_mars_fit"                 "bag_mars_modes"              
#>   [5] "bag_mars_pkgs"                "bag_mars_predict"            
#>   [7] "bag_tree"                     "bag_tree_args"               
#>   [9] "bag_tree_fit"                 "bag_tree_modes"              
#>  [11] "bag_tree_pkgs"                "bag_tree_predict"            
#>  [13] "bart"                         "bart_args"                   
#>  [15] "bart_encoding"                "bart_fit"                    
#>  [17] "bart_modes"                   "bart_pkgs"                   
#>  [19] "bart_predict"                 "boost_tree"                  
#>  [21] "boost_tree_args"              "boost_tree_encoding"         
#>  [23] "boost_tree_fit"               "boost_tree_modes"            
#>  [25] "boost_tree_pkgs"              "boost_tree_predict"          
#>  [27] "C5_rules"                     "C5_rules_args"               
#>  [29] "C5_rules_fit"                 "C5_rules_modes"              
#>  [31] "C5_rules_pkgs"                "C5_rules_predict"            
#>  [33] "cubist_rules"                 "cubist_rules_args"           
#>  [35] "cubist_rules_fit"             "cubist_rules_modes"          
#>  [37] "cubist_rules_pkgs"            "cubist_rules_predict"        
#>  [39] "decision_tree"                "decision_tree_args"          
#>  [41] "decision_tree_encoding"       "decision_tree_fit"           
#>  [43] "decision_tree_modes"          "decision_tree_pkgs"          
#>  [45] "decision_tree_predict"        "discrim_flexible"            
#>  [47] "discrim_flexible_args"        "discrim_flexible_fit"        
#>  [49] "discrim_flexible_modes"       "discrim_flexible_pkgs"       
#>  [51] "discrim_flexible_predict"     "discrim_linear"              
#>  [53] "discrim_linear_args"          "discrim_linear_fit"          
#>  [55] "discrim_linear_modes"         "discrim_linear_pkgs"         
#>  [57] "discrim_linear_predict"       "discrim_quad"                
#>  [59] "discrim_quad_args"            "discrim_quad_fit"            
#>  [61] "discrim_quad_modes"           "discrim_quad_pkgs"           
#>  [63] "discrim_quad_predict"         "discrim_regularized"         
#>  [65] "discrim_regularized_args"     "discrim_regularized_fit"     
#>  [67] "discrim_regularized_modes"    "discrim_regularized_pkgs"    
#>  [69] "discrim_regularized_predict"  "gen_additive_mod"            
#>  [71] "gen_additive_mod_args"        "gen_additive_mod_encoding"   
#>  [73] "gen_additive_mod_fit"         "gen_additive_mod_modes"      
#>  [75] "gen_additive_mod_pkgs"        "gen_additive_mod_predict"    
#>  [77] "linear_reg"                   "linear_reg_args"             
#>  [79] "linear_reg_encoding"          "linear_reg_fit"              
#>  [81] "linear_reg_modes"             "linear_reg_pkgs"             
#>  [83] "linear_reg_predict"           "logistic_reg"                
#>  [85] "logistic_reg_args"            "logistic_reg_encoding"       
#>  [87] "logistic_reg_fit"             "logistic_reg_modes"          
#>  [89] "logistic_reg_pkgs"            "logistic_reg_predict"        
#>  [91] "mars"                         "mars_args"                   
#>  [93] "mars_encoding"                "mars_fit"                    
#>  [95] "mars_modes"                   "mars_pkgs"                   
#>  [97] "mars_predict"                 "mlp"                         
#>  [99] "mlp_args"                     "mlp_encoding"                
#> [101] "mlp_fit"                      "mlp_modes"                   
#> [103] "mlp_pkgs"                     "mlp_predict"                 
#> [105] "models"                       "modes"                       
#> [107] "multinom_reg"                 "multinom_reg_args"           
#> [109] "multinom_reg_encoding"        "multinom_reg_fit"            
#> [111] "multinom_reg_modes"           "multinom_reg_pkgs"           
#> [113] "multinom_reg_predict"         "naive_Bayes"                 
#> [115] "naive_Bayes_args"             "naive_Bayes_fit"             
#> [117] "naive_Bayes_modes"            "naive_Bayes_pkgs"            
#> [119] "naive_Bayes_predict"          "nearest_neighbor"            
#> [121] "nearest_neighbor_args"        "nearest_neighbor_encoding"   
#> [123] "nearest_neighbor_fit"         "nearest_neighbor_modes"      
#> [125] "nearest_neighbor_pkgs"        "nearest_neighbor_predict"    
#> [127] "null_model"                   "null_model_args"             
#> [129] "null_model_encoding"          "null_model_fit"              
#> [131] "null_model_modes"             "null_model_pkgs"             
#> [133] "null_model_predict"           "pls"                         
#> [135] "pls_args"                     "pls_fit"                     
#> [137] "pls_modes"                    "pls_pkgs"                    
#> [139] "pls_predict"                  "poisson_reg"                 
#> [141] "poisson_reg_args"             "poisson_reg_fit"             
#> [143] "poisson_reg_modes"            "poisson_reg_pkgs"            
#> [145] "poisson_reg_predict"          "proportional_hazards"        
#> [147] "proportional_hazards_args"    "proportional_hazards_fit"    
#> [149] "proportional_hazards_modes"   "proportional_hazards_pkgs"   
#> [151] "proportional_hazards_predict" "rand_forest"                 
#> [153] "rand_forest_args"             "rand_forest_encoding"        
#> [155] "rand_forest_fit"              "rand_forest_modes"           
#> [157] "rand_forest_pkgs"             "rand_forest_predict"         
#> [159] "rule_fit"                     "rule_fit_args"               
#> [161] "rule_fit_fit"                 "rule_fit_modes"              
#> [163] "rule_fit_pkgs"                "rule_fit_predict"            
#> [165] "surv_reg"                     "surv_reg_args"               
#> [167] "surv_reg_encoding"            "surv_reg_fit"                
#> [169] "surv_reg_modes"               "surv_reg_pkgs"               
#> [171] "surv_reg_predict"             "survival_reg"                
#> [173] "survival_reg_args"            "survival_reg_fit"            
#> [175] "survival_reg_modes"           "survival_reg_pkgs"           
#> [177] "survival_reg_predict"         "svm_linear"                  
#> [179] "svm_linear_args"              "svm_linear_encoding"         
#> [181] "svm_linear_fit"               "svm_linear_modes"            
#> [183] "svm_linear_pkgs"              "svm_linear_predict"          
#> [185] "svm_poly"                     "svm_poly_args"               
#> [187] "svm_poly_encoding"            "svm_poly_fit"                
#> [189] "svm_poly_modes"               "svm_poly_pkgs"               
#> [191] "svm_poly_predict"             "svm_rbf"                     
#> [193] "svm_rbf_args"                 "svm_rbf_encoding"            
#> [195] "svm_rbf_fit"                  "svm_rbf_modes"               
#> [197] "svm_rbf_pkgs"                 "svm_rbf_predict"