General Interface for Linear Regression Models

linear_reg is a way to generate a specification of a model before fitting and allows the model to be created using different packages in R, Stan, or via Spark. The main arguments for the model are:

penalty: The total amount of regularization in the model. Note that this must be zero for some engines .

mixture: The proportion of L1 regularization in the model. Note that this will be ignored for some engines.

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using the others argument. If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update can be used in lieu of recreating the object from scratch.

linear_reg(mode = "regression", penalty = NULL, mixture = NULL,
  others = list(), ...)

# S3 method for linear_reg
update(object, penalty = NULL, mixture = NULL,
  others = list(), fresh = FALSE, ...)

Arguments

mode
A single character string for the type of model. The only possible value for this model is "regression".

penalty
An non-negative number representing the total amount of regularization (glmnet and spark only).

mixture
A number between zero and one (inclusive) that represents the proportion of regularization that is used for the L2 penalty (i.e. weight decay, or ridge regression) versus L1 (the lasso) (glmnet and spark only).

others
A named list of arguments to be used by the underlying models (e.g., stats::lm, rstanarm::stan_glm, etc.). These are not evaluated until the model is fit and will be substituted into the model fit expression.

...
Used for S3 method consistency. Any arguments passed to the ellipses will result in an error. Use others instead.

object
A linear regression model specification.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

Value

An updated model specification.

Details

The data given to the function are not saved and are only used to determine the mode of the model. For linear_reg, the mode will always be "regression".

The model can be created using the fit() function using the following engines:

R: "lm" or "glmnet"

Stan: "stan"

Spark: "spark"

Engines may have pre-set default arguments when executing the model fit call. These can be changed by using the others argument to pass in the preferred values. For this type of model, the template of the fit calls are:

lm

stats::lm(formula = missing_arg(), data = missing_arg(), weights = missing_arg())

glmnet

glmnet::glmnet(x = missing_arg(), y = missing_arg(), weights = missing_arg(), 
    family = "gaussian")

stan

rstanarm::stan_glm(formula = missing_arg(), data = missing_arg(), 
    weights = missing_arg(), family = "gaussian")

spark

sparklyr::ml_linear_regression(x = missing_arg(), formula = missing_arg(), 
    weight_col = missing_arg())

When using glmnet models, there is the option to pass multiple values (or no values) to the penalty argument. This can have an effect on the model object results. When using the predict method in these cases, the return object type depends on the value of penalty. If a single value is given, the results will be a simple numeric vector. When multiple values or no values for penalty are used in linear_reg, the predict method will return a data frame with columns values and lambda.

For prediction, the stan engine can compute posterior intervals analogous to confidence and prediction intervals. In these instances, the units are the original outcome and when std_error = TRUE, the standard deviation of the posterior distribution (or posterior predictive distribution as appropriate) is returned.

Note

For models created using the spark engine, there are several differences to consider. First, only the formula interface to via fit is available; using fit_xy will generate an error. Second, the predictions will always be in a spark table format. The names will be the same as documented but without the dots. Third, there is no equivalent to factor columns in spark tables so class predictions are returned as character columns. Fourth, to retain the model object for a new R session (via save), the model$fit element of the parsnip object should be serialized via ml_save(object$fit) and separately saved to disk. In a new session, the object can be reloaded and reattached to the parsnip object.

See also

Examples

linear_reg()

#> Linear Regression Model Specification (regression)
#> 

# Parameters can be represented by a placeholder:
linear_reg(penalty = varying())

#> Linear Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   penalty = varying()
#> 

model <- linear_reg(penalty = 10, mixture = 0.1)
model

#> Linear Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   penalty = 10
#>   mixture = 0.1
#> 

update(model, penalty = 1)

#> Linear Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   penalty = 1
#>   mixture = 0.1
#> 

update(model, penalty = 1, fresh = TRUE)

#> Linear Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   penalty = 1
#>