Model predictions across many sub-models

For some models, predictions can be made on sub-models in the model object.

multi_predict(object, ...)

# S3 method for default
multi_predict(object, ...)

Arguments

object
A model_fit object.

...
Optional arguments to pass to predict.model_fit(type = "raw") such as type.

Value

A tibble with the same number of rows as the data being predicted. Mostly likely, there is a list-column named .pred that is a tibble with multiple rows per sub-model.