General Interface for Multinomial Regression Models

multinom_reg() is a way to generate a specification of a model before fitting and allows the model to be created using different packages in R, keras, or Spark. The main arguments for the model are:

penalty: The total amount of regularization in the model. Note that this must be zero for some engines.

mixture: The mixture amounts of different types of regularization (see below). Note that this will be ignored for some engines.

These arguments are converted to their specific names at the time that the model is fit. Other options and arguments can be set using set_engine(). If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

multinom_reg(mode = "classification", penalty = NULL, mixture = NULL)

# S3 method for multinom_reg
update(
  object,
  parameters = NULL,
  penalty = NULL,
  mixture = NULL,
  fresh = FALSE,
  ...
)

Arguments

mode
A single character string for the type of model. The only possible value for this model is "classification".

penalty
A non-negative number representing the total amount of regularization (glmnet, keras, and spark only). For keras models, this corresponds to purely L2 regularization (aka weight decay) while the other models can be a combination of L1 and L2 (depending on the value of mixture).

mixture
A number between zero and one (inclusive) that is the proportion of L1 regularization (i.e. lasso) in the model. When mixture = 1, it is a pure lasso model while mixture = 0 indicates that ridge regression is being used. (glmnet and spark only).

object
A multinomial regression model specification.

parameters
A 1-row tibble or named list with main parameters to update. If the individual arguments are used, these will supersede the values in parameters. Also, using engine arguments in this object will result in an error.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

...
Not used for update().

Details

For multinom_reg(), the mode will always be "classification".

The model can be created using the fit() function using the following engines:

R: "glmnet" (the default), "nnet"

Spark: "spark"

keras: "keras"

Note

For models created using the spark engine, there are several differences to consider. First, only the formula interface to via fit() is available; using fit_xy() will generate an error. Second, the predictions will always be in a spark table format. The names will be the same as documented but without the dots. Third, there is no equivalent to factor columns in spark tables so class predictions are returned as character columns. Fourth, to retain the model object for a new R session (via save()), the model$fit element of the parsnip object should be serialized via ml_save(object$fit) and separately saved to disk. In a new session, the object can be reloaded and reattached to the parsnip object.

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are below.

glmnet

multinom_reg() %>% 
  set_engine("glmnet") %>% 
  set_mode("classification") %>% 
  translate()

## Multinomial Regression Model Specification (classification)
## 
## Computational engine: glmnet 
## 
## Model fit template:
## glmnet::glmnet(x = missing_arg(), y = missing_arg(), weights = missing_arg(), 
##     family = "multinomial")

For glmnet models, the full regularization path is always fit regardless of the value given to penalty. Also, there is the option to pass multiple values (or no values) to the penalty argument. When using the predict() method in these cases, the return value depends on the value of penalty. When using predict(), only a single value of the penalty can be used. When predicting on multiple penalties, the multi_predict() function can be used. It returns a tibble with a list column called .pred that contains a tibble with all of the penalty results.

nnet

multinom_reg() %>% 
  set_engine("nnet") %>% 
  set_mode("classification") %>% 
  translate()

## Multinomial Regression Model Specification (classification)
## 
## Computational engine: nnet 
## 
## Model fit template:
## nnet::multinom(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg(), trace = FALSE)

spark

multinom_reg() %>% 
  set_engine("spark") %>% 
  set_mode("classification") %>% 
  translate()

## Multinomial Regression Model Specification (classification)
## 
## Computational engine: spark 
## 
## Model fit template:
## sparklyr::ml_logistic_regression(x = missing_arg(), formula = missing_arg(), 
##     weight_col = missing_arg(), family = "multinomial")

keras

multinom_reg() %>% 
  set_engine("keras") %>% 
  set_mode("classification") %>% 
  translate()

## Multinomial Regression Model Specification (classification)
## 
## Computational engine: keras 
## 
## Model fit template:
## parsnip::keras_mlp(x = missing_arg(), y = missing_arg(), hidden_units = 1, 
##     act = "linear")

Parameter translations

The standardized parameter names in parsnip can be mapped to their original names in each engine that has main parameters. Each engine typically has a different default value (shown in parentheses) for each parameter.

parsnip	glmnet	spark	keras	nnet

penalty	lambda	reg_param (0)	penalty (0)	decay (0)

mixture	alpha (1)	elastic_net_param (0)	NA	NA

See also

Examples

show_engines("multinom_reg")

#> # A tibble: 4 x 2
#>   engine mode          
#>   <chr>  <chr>         
#> 1 glmnet classification
#> 2 spark  classification
#> 3 keras  classification
#> 4 nnet   classification

multinom_reg()

#> Multinomial Regression Model Specification (classification)
#> 

# Parameters can be represented by a placeholder:
multinom_reg(penalty = varying())

#> Multinomial Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = varying()
#> 

model <- multinom_reg(penalty = 10, mixture = 0.1)
model

#> Multinomial Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = 10
#>   mixture = 0.1
#> 

update(model, penalty = 1)

#> Multinomial Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = 1
#>   mixture = 0.1
#> 

update(model, penalty = 1, fresh = TRUE)

#> Multinomial Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = 1
#>