General Interface for K-Nearest Neighbor Models

nearest_neighbor() is a way to generate a specification of a model before fitting and allows the model to be created using different packages in R. The main arguments for the model are:

neighbors: The number of neighbors considered at each prediction.

weight_func: The type of kernel function that weights the distances between samples.

dist_power: The parameter used when calculating the Minkowski distance. This corresponds to the Manhattan distance with dist_power = 1 and the Euclidean distance with dist_power = 2.

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using set_engine. If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

nearest_neighbor(mode = "unknown", neighbors = NULL,
  weight_func = NULL, dist_power = NULL)

Arguments

mode
A single character string for the type of model. Possible values for this model are "unknown", "regression", or "classification".

neighbors
A single integer for the number of neighbors to consider (often called k).

weight_func
A single character for the type of kernel function used to weight distances between samples. Valid choices are: "rectangular", "triangular", "epanechnikov", "biweight", "triweight", "cos", "inv", "gaussian", "rank", or "optimal".

dist_power
A single number for the parameter used in calculating Minkowski distance.

Details

The model can be created using the fit() function using the following engines:

R: "kknn"

Note

For kknn, the underlying modeling function used is a restricted version of train.kknn() and not kknn(). It is set up in this way so that parsnip can utilize the underlying predict.train.kknn method to predict on new data. This also means that a single value of that function's kernel argument (a.k.a weight_func here) can be supplied

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are:

kknn (classification or regression)

kknn::train.kknn(formula = missing_arg(), data = missing_arg(), 
    kmax = missing_arg())

See also

Examples

nearest_neighbor(neighbors = 11)

#> K-Nearest Neighbor Model Specification (unknown)
#> 
#> Main Arguments:
#>   neighbors = 11
#>