General Interface for null models

null_model() is a way to generate a specification of a model before fitting and allows the model to be created using R. It doens't have any main arguments.

null_model(mode = "classification")

Arguments

mode
A single character string for the type of model. Possible values for this model are "unknown", "regression", or "classification".

Details

The model can be created using the fit() function using the following engines:

R: "parsnip"

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are:

parsnip classification

nullmodel(x = missing_arg(), y = missing_arg())

parsnip regression

nullmodel(x = missing_arg(), y = missing_arg())

See also

Examples

null_model(mode = "regression")

#> Model Specification (regression)
#>