Other predict methods.

These are internal functions not meant to be directly called by the user.

Usage

# S3 method for model_fit
predict_class(object, new_data, ...)

# S3 method for model_fit
predict_classprob(object, new_data, ...)

# S3 method for model_fit
predict_hazard(object, new_data, time, ...)

# S3 method for model_fit
predict_confint(object, new_data, level = 0.95, std_error = FALSE, ...)

# S3 method for model_fit
predict_linear_pred(object, new_data, ...)

predict_linear_pred(object, ...)

# S3 method for model_fit
predict_numeric(object, new_data, ...)

predict_numeric(object, ...)

# S3 method for model_fit
predict_quantile(
  object,
  new_data,
  quantile = (1:9)/10,
  interval = "none",
  level = 0.95,
  ...
)

# S3 method for model_fit
predict_survival(object, new_data, time, interval = "none", level = 0.95, ...)

predict_survival(object, ...)

# S3 method for model_fit
predict_time(object, new_data, ...)

predict_time(object, ...)

Arguments

object

An object of class model_fit

new_data

A rectangular data object, such as a data frame.

...

Arguments to the underlying model's prediction function cannot be passed here (see opts). There are some parsnip related options that can be passed, depending on the value of type. Possible arguments are:

interval: for types of "survival" and "quantile", should interval estimates be added, if available? Options are "none" and "confidence".

level: for types of "conf_int", "pred_int", and "survival" this is the parameter for the tail area of the intervals (e.g. confidence level for confidence intervals). Default value is 0.95.

std_error: add the standard error of fit or prediction (on the scale of the linear predictors) for types of "conf_int" and "pred_int". Default value is FALSE.

quantile: the quantile(s) for quantile regression (not implemented yet)

time: the time(s) for hazard and survival probability estimates.

level

A single numeric value between zero and one for the interval estimates.

std_error

A single logical for whether the standard error should be returned (assuming that the model can compute it).

quantile

A vector of numbers between 0 and 1 for the quantile being predicted.