Model predictions

Apply a model to create different types of predictions. predict can be used for all types of models and used the "type" argument for more specificity. Other functions can be used for each type of prediction (see Details).

# S3 method for model_fit
predict(object, newdata, type = NULL, ...)

# S3 method for model_fit
predict_class(object, newdata, ...)

predict_class(object, ...)

# S3 method for model_fit
predict_classprob(object, newdata, ...)

predict_classprob(object, ...)

# S3 method for model_fit
predict_num(object, newdata, ...)

predict_num(object, ...)

Arguments

object
An object of class model_fit

newdata
A rectangular data object, such as a data frame.

type
A single character value or NULL. Possible values are "numeric", "class", "probs", "link", or "raw" (the last two are not yet implemented). When NULL, predict will choose an appropriate value based on the model's mode.

...
Arguments to pass to other methods (not currently used).

Value

See Details.

Details

If "type" is not supplied to predict, then a choice is made (type = "numeric" for regression models and type = "class" for classification).

predict is designed to provide a tidy results. There will be a tibble as many rows in the output as there are rows in newdata and the column names will be predictable. For numeric results with a single outcome, the tibble will have a .pred column and .pred_Yname for multivariate results. For hard class predictions, the column is named .pred_class and, when type = "prob", the columns are .pred_classlevel.

The more specific prediction functions (e.g. predict_num) return non-tibble results. predict_num generates a vector (for univariate models) or a data frame (multivariate). predict_class returns a factor and predict_classprob returns a data frame with columns for the factor levels.

Examples

library(dplyr)

lm_model <-
  linear_reg() %>%
  fit(mpg ~ ., data = mtcars %>% slice(11:32), engine = "lm")

predict(lm_model, mtcars %>% slice(1:10) %>% select(-mpg))

#> # A tibble: 10 x 1
#>    .pred
#>    <dbl>
#>  1  23.4
#>  2  23.3
#>  3  27.6
#>  4  21.5
#>  5  17.6
#>  6  21.6
#>  7  13.9
#>  8  21.7
#>  9  25.6
#> 10  17.1

predict_num(lm_model, mtcars %>% slice(1:10) %>% select(-mpg))

#>  [1] 23.43006 23.29639 27.62870 21.54678 17.57028 21.58316 13.90725 21.70176
#>  [9] 25.61122 17.11206