Model Predictions

Apply a model to create different types of predictions. predict is used when the outcome is a simple number. For classification models, predict_class and predict_classprob can be used.

# S3 method for model_fit
predict(object, newdata, ...)

# S3 method for model_fit
predict_class(object, newdata, ...)

predict_class(object, ...)

# S3 method for model_fit
predict_classprob(object, newdata, ...)

predict_classprob(object, ...)

Arguments

object
An object of class model_fit

newdata
A rectangular data object, such as a data frame.

...
Arguments to pass to other methods (not currently used).

Value

predict returns an unnamed numeric vector. predict_class returns a factor variable and predict_classprob generates a tibble with a column for each class level.