Model predictions

Apply a model to create different types of predictions. predict can be used for all types of models and used the "type" argument for more specificity. Other functions can be used for each type of prediction (see Details).

# S3 method for model_fit
predict(object, new_data, type = NULL,
  opts = list(), ...)

# S3 method for model_fit
predict_class(object, new_data, ...)

predict_class(object, ...)

# S3 method for model_fit
predict_classprob(object, new_data, ...)

predict_classprob(object, ...)

# S3 method for model_fit
predict_confint(object, new_data, level = 0.95,
  std_error = FALSE, ...)

predict_confint(object, ...)

# S3 method for model_fit
predict_predint(object, new_data, level = 0.95,
  std_error = FALSE, ...)

predict_predint(object, ...)

# S3 method for model_fit
predict_num(object, new_data, ...)

predict_num(object, ...)

# S3 method for model_fit
predict_raw(object, new_data, opts = list(), ...)

predict_raw(object, ...)

Arguments

object
An object of class model_fit

new_data
A rectangular data object, such as a data frame.

type
A single character value or NULL. Possible values are "numeric", "class", "probs", "conf_int", "pred_int", or "raw". When NULL, predict will choose an appropriate value based on the model's mode.

opts
A list of optional arguments to the underlying predict function that will be used when type = "raw". The list should not include options for the model object or the new data being predicted.

...
Ignored. To pass arguments to pass to the underlying function when predict_raw or predict.model_fit(type = "raw"), use the opts argument.

level
A single numeric value between zero and one for the interval estimates.

std_error
A single logical for wether the standard error should be returned (assuming that the model can compute it).

Value

With the exception of type = "raw", the results of predict.model_fit will be a tibble as many rows in the output as there are rows in new_data and the column names will be predictable. For numeric results with a single outcome, the tibble will have a .pred column and .pred_Yname for multivariate results. For hard class predictions, the column is named .pred_class and, when type = "prob", the columns are .pred_classlevel. type = "conf_int" and type = "pred_int"return tibbles with columns.pred_lowerand.pred_upperwith an attribute for the confidence level. In the case where intervals can be produces for class probabilities (or other non-scalar outputs), the columns will be named.pred_lower_classlevel` and so on.

Using type = "raw" with predict.model_fit (or using predict_raw) will return the unadulterated results of the prediction function.

The more specific prediction functions (e.g. predict_num) can return non-tibble results. predict_num generates a vector (for univariate models) or a data frame (multivariate). predict_class returns a factor and predict_classprob returns a data frame with columns for the factor levels. Other type-specific prediction functions return tibbles if it is natural to produce more than one column of output.

In the case of Spark-based models, since table columns cannot contain dots, the same convention is used except 1) no dots appear in names and 2) vectors are never returned but type-specific prediction functions.

Details

If "type" is not supplied to predict, then a choice is made (type = "numeric" for regression models and type = "class" for classification).

predict is designed to provide a tidy result (see "Value" section below) in a tibble output format.

Examples

library(dplyr)

lm_model <-
  linear_reg() %>%
  fit(mpg ~ ., data = mtcars %>% slice(11:32), engine = "lm")

predict(lm_model, mtcars %>% slice(1:10) %>% select(-mpg))

#> # A tibble: 10 x 1
#>    .pred
#>    <dbl>
#>  1  23.4
#>  2  23.3
#>  3  27.6
#>  4  21.5
#>  5  17.6
#>  6  21.6
#>  7  13.9
#>  8  21.7
#>  9  25.6
#> 10  17.1

predict_num(lm_model, mtcars %>% slice(1:10) %>% select(-mpg))

#>  [1] 23.43006 23.29639 27.62870 21.54678 17.57028 21.58316 13.90725 21.70176
#>  [9] 25.61122 17.11206

predict(lm_model, mtcars %>% slice(1:10) %>% select(-mpg), type = "conf_int")

#> # A tibble: 10 x 2
#>    .pred_lower .pred_upper
#>          <dbl>       <dbl>
#>  1       16.6         30.2
#>  2       16.9         29.6
#>  3       23.2         32.1
#>  4       16.6         26.5
#>  5       13.6         21.5
#>  6       16.0         27.2
#>  7        8.69        19.1
#>  8       15.0         28.4
#>  9       11.7         39.6
#> 10       10.3         23.9

predict(
   lm_model,
   mtcars %>% slice(1:3) %>% select(-mpg),
   type = "raw",
   opts = list(type = "terms"))

#>            cyl       disp        hp        drat       wt      qsec         vs
#> 1 -0.001433177 -0.8113275 0.6303467 -0.06120265 2.413981 -1.567729  0.2006406
#> 2 -0.001433177 -0.8113275 0.6303467 -0.06120265 1.448871 -0.736286  0.2006406
#> 3 -0.009315653 -1.3336453 0.8557288 -0.05014798 3.549406  1.624418 -0.3511210
#>        am      gear       carb
#> 1 2.88774 0.0251268 -0.2497240
#> 2 2.88774 0.0251268 -0.2497240
#> 3 2.88774 0.0251268  0.4668753
#> attr(,"constant")
#> [1] 19.96364