Model predictions

Apply a model to create different types of predictions. predict can be used for all types of models and used the "type" argument for more specificity.

# S3 method for model_fit
predict(object, new_data, type = NULL,
  opts = list(), ...)

# S3 method for model_fit
predict_raw(object, new_data, opts = list(), ...)

predict_raw(object, ...)

Arguments

object
An object of class model_fit

new_data
A rectangular data object, such as a data frame.

type
A single character value or NULL. Possible values are "numeric", "class", "probs", "conf_int", "pred_int", "quantile", or "raw". When NULL, predict will choose an appropriate value based on the model's mode.

opts
A list of optional arguments to the underlying predict function that will be used when type = "raw". The list should not include options for the model object or the new data being predicted.

...
Ignored. To pass arguments to pass to the underlying function when predict.model_fit(type = "raw"), use the opts argument.

Value

With the exception of type = "raw", the results of predict.model_fit will be a tibble as many rows in the output as there are rows in new_data and the column names will be predictable.

For numeric results with a single outcome, the tibble will have a .pred column and .pred_Yname for multivariate results.

For hard class predictions, the column is named .pred_class and, when type = "prob", the columns are .pred_classlevel.

type = "conf_int" and type = "pred_int" return tibbles with columns .pred_lower and .pred_upper with an attribute for the confidence level. In the case where intervals can be produces for class probabilities (or other non-scalar outputs), the columns will be named .pred_lower_classlevel and so on.

Quantile predictions return a tibble with a column .pred, which is a list-column. Each list element contains a tibble with columns .pred and .quantile (and perhaps other columns).

Using type = "raw" with predict.model_fit (or using predict_raw) will return the unadulterated results of the prediction function.

In the case of Spark-based models, since table columns cannot contain dots, the same convention is used except 1) no dots appear in names and 2) vectors are never returned but type-specific prediction functions.

Details

If "type" is not supplied to predict, then a choice is made (type = "numeric" for regression models and type = "class" for classification).

predict is designed to provide a tidy result (see "Value" section below) in a tibble output format.

When using type = "conf_int" and type = "pred_int", the options level and std_error can be used. The latter is a logical for an extra column of standard error values (if available).

Examples

library(dplyr)

lm_model <-
  linear_reg() %>%
  set_engine("lm") %>%
  fit(mpg ~ ., data = mtcars %>% slice(11:32))

pred_cars <-
  mtcars %>%
  slice(1:10) %>%
  select(-mpg)

predict(lm_model, pred_cars)

#> [38;5;246m# A tibble: 10 x 1[39m
#>    .pred
#>    [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m  23.4
#> [38;5;250m 2[39m  23.3
#> [38;5;250m 3[39m  27.6
#> [38;5;250m 4[39m  21.5
#> [38;5;250m 5[39m  17.6
#> [38;5;250m 6[39m  21.6
#> [38;5;250m 7[39m  13.9
#> [38;5;250m 8[39m  21.7
#> [38;5;250m 9[39m  25.6
#> [38;5;250m10[39m  17.1

predict(
  lm_model,
  pred_cars,
  type = "conf_int",
  level = 0.90
)

#> [38;5;246m# A tibble: 10 x 2[39m
#>    .pred_lower .pred_upper
#>          [3m[38;5;246m<dbl>[39m[23m       [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m       17.9         29.0
#> [38;5;250m 2[39m       18.1         28.5
#> [38;5;250m 3[39m       24.0         31.3
#> [38;5;250m 4[39m       17.5         25.6
#> [38;5;250m 5[39m       14.3         20.8
#> [38;5;250m 6[39m       17.0         26.2
#> [38;5;250m 7[39m        9.65        18.2
#> [38;5;250m 8[39m       16.2         27.2
#> [38;5;250m 9[39m       14.2         37.0
#> [38;5;250m10[39m       11.5         22.7

predict(
  lm_model,
  pred_cars,
  type = "raw",
  opts = list(type = "terms")
)

#>             cyl       disp         hp        drat         wt      qsec
#> 1  -0.001433177 -0.8113275  0.6303467 -0.06120265  2.4139815 -1.567729
#> 2  -0.001433177 -0.8113275  0.6303467 -0.06120265  1.4488706 -0.736286
#> 3  -0.009315653 -1.3336453  0.8557288 -0.05014798  3.5494061  1.624418
#> 4  -0.001433177  0.1730406  0.6303467  0.12009386  0.1620561  2.856736
#> 5   0.006449298  1.1975870 -0.2314083  0.10461733 -0.6895124 -0.736286
#> 6  -0.001433177 -0.1584303  0.6966356  0.19084372 -0.7652074  4.014817
#> 7   0.006449298  1.1975870 -1.1594522  0.09135173 -1.1815297 -2.488255
#> 8  -0.009315653 -0.9449204  1.2667197 -0.01477305  0.2566748  3.688179
#> 9  -0.009315653 -1.0041833  0.8292133 -0.06562451  0.4080647  7.993866
#> 10 -0.001433177 -0.7349888  0.4579957 -0.06562451 -0.6895124  1.164155
#>            vs       am        gear       carb
#> 1   0.2006406  2.88774  0.02512680 -0.2497240
#> 2   0.2006406  2.88774  0.02512680 -0.2497240
#> 3  -0.3511210  2.88774  0.02512680  0.4668753
#> 4  -0.3511210 -2.40645 -0.06700481  0.4668753
#> 5   0.2006406 -2.40645 -0.06700481  0.2280089
#> 6  -0.3511210 -2.40645 -0.06700481  0.4668753
#> 7   0.2006406 -2.40645 -0.06700481 -0.2497240
#> 8  -0.3511210 -2.40645  0.02512680  0.2280089
#> 9  -0.3511210 -2.40645  0.02512680  0.2280089
#> 10 -0.3511210 -2.40645  0.02512680 -0.2497240
#> attr(,"constant")
#> [1] 19.96364