Declare a computational engine and specific arguments

set_engine() is used to specify which package or system will be used to fit the model, along with any arguments specific to that software.

Usage

set_engine(object, engine, ...)

Arguments

object

A model specification.

engine

A character string for the software that should be used to fit the model. This is highly dependent on the type of model (e.g. linear regression, random forest, etc.).

...

Any optional arguments associated with the chosen computational engine. These are captured as quosures and can be varying().

Value

An updated model specification.

Engines

Based on the currently loaded packages, the following lists the set of engines available to each model specification.

bag_mars(): No engines currently available

bag_tree(): No engines currently available

bart(): dbarts

boost_tree(): C5.0, spark, xgboost

C5_rules(): No engines currently available

cubist_rules(): No engines currently available

decision_tree(): C5.0, rpart, spark

discrim_flexible(): No engines currently available

discrim_linear(): No engines currently available

discrim_quad(): No engines currently available

discrim_regularized(): No engines currently available

gen_additive_mod(): mgcv

linear_reg(): brulee, glm, glmnet, keras, lm, spark, stan

logistic_reg(): LiblineaR, brulee, glm, glmnet, keras, spark, stan

mars(): earth

mlp(): brulee, keras, nnet

multinom_reg(): brulee, glmnet, keras, nnet, spark

naive_Bayes(): No engines currently available

nearest_neighbor(): kknn

null_model(): parsnip

pls(): No engines currently available

poisson_reg(): No engines currently available

proportional_hazards(): No engines currently available

rand_forest(): randomForest, ranger, spark

rule_fit(): No engines currently available

surv_reg(): flexsurv, survival

survival_reg(): No engines currently available

svm_linear(): LiblineaR, kernlab

svm_poly(): kernlab

svm_rbf(): kernlab, liquidSVM

Examples

# First, set general arguments using the standardized names
mod <-
  logistic_reg(penalty = 0.01, mixture = 1/3) %>%
  # now say how you want to fit the model and another other options
  set_engine("glmnet", nlambda = 10)
translate(mod, engine = "glmnet")
#> Logistic Regression Model Specification (classification)
#> 
#> Main Arguments:
#>   penalty = 0.01
#>   mixture = 1/3
#> 
#> Engine-Specific Arguments:
#>   nlambda = 10
#> 
#> Computational engine: glmnet 
#> 
#> Model fit template:
#> glmnet::glmnet(x = missing_arg(), y = missing_arg(), weights = missing_arg(), 
#>     alpha = 1/3, nlambda = 10, family = "binomial")