Parametric survival regression

This function is deprecated in favor of survival_reg() which uses the "censored regression" mode.

surv_reg() defines a parametric survival model.

More information on how parsnip is used for modeling is at https://www.tidymodels.org/.

Usage

surv_reg(mode = "regression", engine = "survival", dist = NULL)

Arguments

mode

A single character string for the prediction outcome mode. The only possible value for this model is "regression".

engine

A single character string specifying what computational engine to use for fitting.

dist

A character string for the probability distribution of the outcome. The default is "weibull".

Details

This function only defines what type of model is being fit. Once an engine is specified, the method to fit the model is also defined. See set_engine() for more on setting the engine, including how to set engine arguments.

The model is not trained or fit until the fit() function is used with the data.

Each of the arguments in this function other than mode and engine are captured as quosures. To pass values programmatically, use the injection operator like so:

value <- 1
surv_reg(argument = !!value)

Since survival models typically involve censoring (and require the use of survival::Surv() objects), the fit.model_spec() function will require that the survival model be specified via the formula interface.

References

https://www.tidymodels.org, Tidy Modeling with R, searchable table of parsnip models