Determine varying arguments

varying_args takes a model specification and lists all of the arguments along with whether they are fully specified or not.

varying_args(x, id, ...)

# S3 method for model_spec
varying_args(x, id = NULL, ...)

# S3 method for recipe
varying_args(x, id = NULL, ...)

# S3 method for step
varying_args(x, id = NULL, ...)

Arguments

x
An object

id
A string describing the object x.

...
Not currently used.

Value

A tibble with columns for the parameter name (name), whether is contains any varying value (varying), the id for the object, and the class that was used to call the method (type).

Examples

library(dplyr)
library(rlang)

#> 
#> Attaching package: ‘rlang’

#> The following objects are masked from ‘package:testthat’:
#> 
#>     is_false, is_null, is_true

rand_forest() %>% varying_args(id = "plain")

#> Warning: `list_len()` is soft-deprecated as of rlang 0.2.0.
#> Please use `new_list()` instead
#> [90mThis warning is displayed once per session.[39m

#> [38;5;246m# A tibble: 3 x 4[39m
#>   name  varying id    type      
#>   [3m[38;5;246m<chr>[39m[23m [3m[38;5;246m<lgl>[39m[23m   [3m[38;5;246m<chr>[39m[23m [3m[38;5;246m<chr>[39m[23m     
#> [38;5;250m1[39m mtry  FALSE   plain model_spec
#> [38;5;250m2[39m trees FALSE   plain model_spec
#> [38;5;250m3[39m min_n FALSE   plain model_spec

rand_forest(mtry = varying()) %>% varying_args(id = "one arg")

#> [38;5;246m# A tibble: 3 x 4[39m
#>   name  varying id      type      
#>   [3m[38;5;246m<chr>[39m[23m [3m[38;5;246m<lgl>[39m[23m   [3m[38;5;246m<chr>[39m[23m   [3m[38;5;246m<chr>[39m[23m     
#> [38;5;250m1[39m mtry  TRUE    one arg model_spec
#> [38;5;250m2[39m trees FALSE   one arg model_spec
#> [38;5;250m3[39m min_n FALSE   one arg model_spec

rand_forest() %>%
  set_engine("ranger", sample.fraction = varying()) %>%
  varying_args(id = "only eng_args")

#> [38;5;246m# A tibble: 4 x 4[39m
#>   name            varying id            type      
#>   [3m[38;5;246m<chr>[39m[23m           [3m[38;5;246m<lgl>[39m[23m   [3m[38;5;246m<chr>[39m[23m         [3m[38;5;246m<chr>[39m[23m     
#> [38;5;250m1[39m mtry            FALSE   only eng_args model_spec
#> [38;5;250m2[39m trees           FALSE   only eng_args model_spec
#> [38;5;250m3[39m min_n           FALSE   only eng_args model_spec
#> [38;5;250m4[39m sample.fraction TRUE    only eng_args model_spec

rand_forest() %>%
  set_engine(
    "ranger",
    strata = expr(Class),
     sampsize = c(varying(), varying())
  ) %>%
  varying_args(id = "add an expr")

#> [38;5;246m# A tibble: 5 x 4[39m
#>   name     varying id          type      
#>   [3m[38;5;246m<chr>[39m[23m    [3m[38;5;246m<lgl>[39m[23m   [3m[38;5;246m<chr>[39m[23m       [3m[38;5;246m<chr>[39m[23m     
#> [38;5;250m1[39m mtry     FALSE   add an expr model_spec
#> [38;5;250m2[39m trees    FALSE   add an expr model_spec
#> [38;5;250m3[39m min_n    FALSE   add an expr model_spec
#> [38;5;250m4[39m strata   FALSE   add an expr model_spec
#> [38;5;250m5[39m sampsize TRUE    add an expr model_spec

 rand_forest() %>%
   set_engine("ranger", classwt = c(class1 = 1, class2 = varying())) %>%
   varying_args(id = "list of values")

#> [38;5;246m# A tibble: 4 x 4[39m
#>   name    varying id             type      
#>   [3m[38;5;246m<chr>[39m[23m   [3m[38;5;246m<lgl>[39m[23m   [3m[38;5;246m<chr>[39m[23m          [3m[38;5;246m<chr>[39m[23m     
#> [38;5;250m1[39m mtry    FALSE   list of values model_spec
#> [38;5;250m2[39m trees   FALSE   list of values model_spec
#> [38;5;250m3[39m min_n   FALSE   list of values model_spec
#> [38;5;250m4[39m classwt TRUE    list of values model_spec